Fix including JPetRecoImageTools as module
