+++ ImageReconstruction/CMakeLists.txt
@@ -17,6 +17,8 @@ set(projectName ImageReconstruction)
 set(projectBinary ${projectName}.x)
 project(${projectName} CXX)
 
+include_directories(../modules/large_barrel/JPetRecoImageTools/)
+
 file(GLOB HEADERS *.h)
 file(GLOB SOURCES *.cpp)
 file(GLOB MAIN_CPP main.cpp)
@@ -25,6 +27,7 @@ list(REMOVE_ITEM SOURCES ${UNIT_TEST_SOURCES})
 file(GLOB SOURCES_WITHOUT_MAIN *.cpp)
 list(REMOVE_ITEM SOURCES_WITHOUT_MAIN ${UNIT_TEST_SOURCES})
 list(REMOVE_ITEM SOURCES_WITHOUT_MAIN ${MAIN_CPP})
+
 
 find_package(FFTW REQUIRED)
 include_directories(${FFTW_INCLUDE_DIRS})
@@ -75,7 +78,7 @@ include_directories(${Framework_INCLUDE_DIRS})
 add_definitions(${Framework_DEFINITIONS})
 
 add_executable(${projectBinary} ${SOURCES} ${HEADERS})
-target_link_libraries(${projectBinary} JPetFramework ${PNG_LIBRARIES} ${FFTW_LIBRARIES})
+target_link_libraries(${projectBinary} JPetFramework ${PNG_LIBRARIES} ${FFTW_LIBRARIES} JPetRecoImageTools)
 
 add_custom_target(clean_data_${projectName}
   COMMAND rm -f *.tslot.*.root *.phys.*.root *.sig.root
