+++ modules/large_barrel/JPetRecoImageTools/JPetRecoImageTools.cpp
@@ -20,30 +20,24 @@ JPetRecoImageTools::JPetRecoImageTools() {}
 
 JPetRecoImageTools::~JPetRecoImageTools() {}
 
-double
-JPetRecoImageTools::nearestNeighbour(int i, double y,
-                                     std::function< double(int, int) >& func)
+double JPetRecoImageTools::nearestNeighbour(int i, double y, std::function<double(int, int)>& func)
 {
   int j = std::round(y);
   return func(i, j);
 }
 
-double JPetRecoImageTools::linear(int i, double y,
-                                  std::function< double(int, int) >& func)
+double JPetRecoImageTools::linear(int i, double y, std::function<double(int, int)>& func)
 {
   int j = std::floor(y);
   double weight = std::abs(y - std::floor(y));
   return (1 - weight) * func(i, j) + weight * func(i, j + 1);
 }
 
-std::function< double(int, int) >
-JPetRecoImageTools::matrixGetterFactory(const Matrix2D& emissionMatrix,
-                                        bool isTransposed)
+std::function<double(int, int)> JPetRecoImageTools::matrixGetterFactory(const Matrix2D& emissionMatrix, bool isTransposed)
 {
   if (!isTransposed) {
     return [&emissionMatrix](int i, int j) {
-      if (i >= 0 && i < (int)emissionMatrix[0].size() && j >= 0 &&
-          j < (int)emissionMatrix.size()) {
+      if (i >= 0 && i < (int)emissionMatrix[0].size() && j >= 0 && j < (int)emissionMatrix.size()) {
         return emissionMatrix[i][j];
       } else {
         return 0;
@@ -51,8 +45,7 @@ JPetRecoImageTools::matrixGetterFactory(const Matrix2D& emissionMatrix,
     };
   } else {
     return [&emissionMatrix](int i, int j) {
-      if (i >= 0 && i < (int)emissionMatrix.size() && j >= 0 &&
-          j < (int)emissionMatrix[0].size()) {
+      if (i >= 0 && i < (int)emissionMatrix.size() && j >= 0 && j < (int)emissionMatrix[0].size()) {
         return emissionMatrix[j][i];
       } else {
         return 0;
@@ -61,11 +54,11 @@ JPetRecoImageTools::matrixGetterFactory(const Matrix2D& emissionMatrix,
   }
 }
 
-JPetRecoImageTools::Matrix2DProj
-JPetRecoImageTools::createSinogramWithSingleInterpolation(
-  Matrix2D& emissionMatrix, int nViews, int nScans, double angleBeg,
-  double angleEnd, InterpolationFunc interpolationFunction,
-  RescaleFunc rescaleFunc, int rescaleMinCutoff, int rescaleFactor)
+JPetRecoImageTools::Matrix2DProj JPetRecoImageTools::createSinogramWithSingleInterpolation(Matrix2D& emissionMatrix, int nViews, int nScans,
+    double angleBeg, double angleEnd,
+    InterpolationFunc interpolationFunction,
+    RescaleFunc rescaleFunc, int rescaleMinCutoff,
+    int rescaleFactor)
 {
   assert(emissionMatrix.size() > 0);
   assert(emissionMatrix.size() == emissionMatrix[0].size());
@@ -74,7 +67,7 @@ JPetRecoImageTools::createSinogramWithSingleInterpolation(
   assert(angleBeg < angleEnd);
   assert(rescaleMinCutoff < rescaleFactor);
 
-  Matrix2DProj proj(nScans, std::vector< double >(nViews));
+  Matrix2DProj proj(nScans, std::vector<double>(nViews));
 
   float stepsize = (angleEnd - angleBeg) / nViews;
   assert(stepsize > 0);
@@ -83,17 +76,15 @@ JPetRecoImageTools::createSinogramWithSingleInterpolation(
   for (auto phi = angleBeg; phi < angleEnd; phi = phi + stepsize, viewIndex++) {
     for (auto scanNumber = 0; scanNumber < nScans; scanNumber++) {
       proj[nScans - 1 - scanNumber][viewIndex] =
-        JPetRecoImageTools::calculateProjection(
-          emissionMatrix, phi, scanNumber, nScans, interpolationFunction);
+        JPetRecoImageTools::calculateProjection(emissionMatrix, phi, scanNumber, nScans, interpolationFunction);
     }
   }
   rescaleFunc(proj, rescaleMinCutoff, rescaleFactor);
   return proj;
 }
 
-double JPetRecoImageTools::calculateProjection(
-  const Matrix2D& emissionMatrix, double angle, int scanNumber, int nScans,
-  InterpolationFunc& interpolationFunction)
+double JPetRecoImageTools::calculateProjection(const Matrix2D& emissionMatrix, double angle, int scanNumber, int nScans,
+    InterpolationFunc& interpolationFunction)
 {
   int N = scanNumber - std::floor(nScans / 2);
   const int kInputMatrixSize = emissionMatrix.size();
@@ -118,40 +109,35 @@ double JPetRecoImageTools::calculateProjection(
   /// we use x = a* y +b, and we iterate over columns of the matrix (y)
   bool angleRange45To125 = std::abs(sin) > kSin45or125deg;
   double divided = 1.;
-  std::function< double(int, int) > matrixGet;
+  std::function<double(int, int)> matrixGet;
 
   if (angleRange45To125) {
     assert(sin);
     a = -cos / sin;
     b = (N - cos - sin) / sin;
     b *= scale;
-    matrixGet = matrixGetterFactory(
-                  emissionMatrix, false); // The matrix  elements will be taken as (x,y).
+    matrixGet = matrixGetterFactory(emissionMatrix, false); // The matrix  elements will be taken as (x,y).
     divided = std::abs(sin);
   } else {
     assert(cos);
     a = -sin / cos;
     b = (N - cos - sin) / cos;
     b *= scale;
-    matrixGet = matrixGetterFactory(
-                  emissionMatrix,
-                  true); // The matrix  elements will be taken as (y, x) - transposed.
+    matrixGet = matrixGetterFactory(emissionMatrix, true); // The matrix  elements will be taken as (y, x) - transposed.
     divided = std::abs(cos);
   }
   const int kMatrixCenter = std::floor(emissionMatrix.size() / 2);
   double value = 0.;
   for (auto i = -kMatrixCenter; i < kMatrixCenter; i++) {
-    value += interpolationFunction(i + kMatrixCenter, a * i + b + kMatrixCenter,
-                                   matrixGet);
+    value += interpolationFunction(i + kMatrixCenter, a * i + b + kMatrixCenter, matrixGet);
   }
   value /= divided;
   return value;
 }
 
-JPetRecoImageTools::Matrix2DProj
-JPetRecoImageTools::createSinogramWithDoubleInterpolation(
-  Matrix2D& emissionMatrix, int nAngles, RescaleFunc rescaleFunc,
-  int rescaleMinCutoff, int rescaleFactor)
+JPetRecoImageTools::Matrix2DProj JPetRecoImageTools::createSinogramWithDoubleInterpolation(Matrix2D& emissionMatrix, int nAngles,
+    RescaleFunc rescaleFunc, int rescaleMinCutoff,
+    int rescaleFactor)
 {
   assert(emissionMatrix.size() > 0);
   assert(emissionMatrix.size() == emissionMatrix[0].size());
@@ -162,34 +148,27 @@ JPetRecoImageTools::createSinogramWithDoubleInterpolation(
   double center2 = center * center;
   double angleStep = M_PI / nAngles;
 
-  Matrix2DProj proj(imageSize, std::vector< double >(nAngles));
-  std::function< double(int, int) > matrixGet;
-  matrixGet = matrixGetterFactory(
-                emissionMatrix, false); // The matrix  elements will be taken as (x,y).
+  Matrix2DProj proj(imageSize, std::vector<double>(nAngles));
+  std::function<double(int, int)> matrixGet;
+  matrixGet = matrixGetterFactory(emissionMatrix, false); // The matrix  elements will be taken as (x,y).
   for (int angle = 0; angle < nAngles; angle++) {
     double cos = std::cos((double)angle * angleStep);
     double sin = std::sin((double)angle * angleStep);
     for (int scanNumber = 0; scanNumber < imageSize - 1; scanNumber++) {
-      proj[scanNumber][angle] = calculateProjection2(
-                                  scanNumber, cos, sin, imageSize, center, center2, matrixGet);
+      proj[scanNumber][angle] = calculateProjection2(scanNumber, cos, sin, imageSize, center, center2, matrixGet);
     }
   }
   rescaleFunc(proj, rescaleMinCutoff, rescaleFactor);
   return proj;
 }
 
-double JPetRecoImageTools::calculateProjection2(
-  int step, double cos, double sin, int imageSize, double center,
-  double center2, std::function< double(int, int) > matrixGet)
+double JPetRecoImageTools::calculateProjection2(int step, double cos, double sin, int imageSize, double center, double center2,
+    std::function<double(int, int)> matrixGet)
 {
   double stepMinusCenter = step - center;
   double xtmp = center + stepMinusCenter * cos;
   double ytmp = center - stepMinusCenter * sin;
-  int nmin =
-    step == 0 || step == imageSize - 1
-    ? 0
-    : std::floor(center - std::sqrt((center2 - (stepMinusCenter *
-                                     stepMinusCenter))));
+  int nmin = step == 0 || step == imageSize - 1 ? 0 : std::floor(center - std::sqrt((center2 - (stepMinusCenter * stepMinusCenter))));
   double p = 0.0;
   for (int n = nmin; n < imageSize - nmin; n++) {
     double nMinusCenter = (double)n - center;
@@ -220,60 +199,56 @@ void JPetRecoImageTools::rescale(Matrix2DProj& matrix, double minCutoff, double 
 
   /// datamin represents the constant background factor.
   if ((datamax - datamin) == 0.) {
-    WARNING(
-      "Maximum value in the matrix to rescale is 0. No rescaling performed.");
+    WARNING("Maximum value in the matrix to rescale is 0. No rescaling performed.");
     return;
   }
 
   for (unsigned int k = 0; k < matrix.size(); k++) {
     for (unsigned int j = 0; j < matrix[0].size(); j++) {
-      matrix[k][j] =
-        (matrix[k][j] - datamin) * rescaleFactor / (datamax - datamin);
+      matrix[k][j] = (matrix[k][j] - datamin) * rescaleFactor / (datamax - datamin);
     }
   }
 }
 
-JPetRecoImageTools::Matrix2DProj
-JPetRecoImageTools::backProject(Matrix2DProj& sinogram, int nAngles,
-                                int imageSize = sinogram.size();
-                                double center = (double)(imageSize - 1) / 2.0;
-                                double center2 = center* center;
-                                double angleStep = M_PI / (double)nAngles;
-
-                                std::vector< double >(imageSize));
-
-for (int angle = 0; angle < nAngles; angle++)
+JPetRecoImageTools::Matrix2DProj JPetRecoImageTools::backProject(Matrix2DProj& sinogram, int nAngles, RescaleFunc rescaleFunc, int rescaleMinCutoff,
+    int rescaleFactor)
 {
-  double cos = std::cos((double)angle * (double)angleStep);
-  double sin = std::sin((double)angle * (double)angleStep);
+  int imageSize = sinogram.size();
+  double center = (double)(imageSize - 1) / 2.0;
+  double center2 = center * center;
+  double angleStep = M_PI / (double)nAngles;
 
-  for (int x = 0; x < imageSize; x++) {
-    double xMinusCenter = (double)x - center;
-    double xMinusCenter2 = xMinusCenter * xMinusCenter;
-    double ttemp = xMinusCenter * cos + center;
+  Matrix2DProj reconstructedProjection(imageSize, std::vector<double>(imageSize));
 
-    for (int y = 0; y < imageSize; y++) {
-      double yMinusCenter = (double)y - center;
-      double yMinusCenter2 = yMinusCenter * yMinusCenter;
-      if (yMinusCenter2 + xMinusCenter2 < center2) {
-        double t = ttemp - yMinusCenter * sin;
-        int n = std::floor(t + 0.5F);
-        reconstructedProjection[y][x] += sinogram[n][angle];
+  for (int angle = 0; angle < nAngles; angle++) {
+    double cos = std::cos((double)angle * (double)angleStep);
+    double sin = std::sin((double)angle * (double)angleStep);
+
+    for (int x = 0; x < imageSize; x++) {
+      double xMinusCenter = (double)x - center;
+      double xMinusCenter2 = xMinusCenter * xMinusCenter;
+      double ttemp = xMinusCenter * cos + center;
+
+      for (int y = 0; y < imageSize; y++) {
+        double yMinusCenter = (double)y - center;
+        double yMinusCenter2 = yMinusCenter * yMinusCenter;
+        if (yMinusCenter2 + xMinusCenter2 < center2) {
+          double t = ttemp - yMinusCenter * sin;
+          int n = std::floor(t + 0.5F);
+          reconstructedProjection[y][x] += sinogram[n][angle];
+        }
       }
     }
   }
-}
 
-for (int x = 0; x < imageSize; x++)
-{
-  for (int y = 0; y < imageSize; y++) {
-    reconstructedProjection[y][x] *= angleStep;
+  for (int x = 0; x < imageSize; x++) {
+    for (int y = 0; y < imageSize; y++) {
+      reconstructedProjection[y][x] *= angleStep;
+    }
   }
+  rescaleFunc(reconstructedProjection, rescaleMinCutoff, rescaleFactor);
+  return reconstructedProjection;
 }
-rescaleFunc(reconstructedProjection, rescaleMinCutoff, rescaleFactor);
-return reconstructedProjection;
-}
-
 
 JPetRecoImageTools::Matrix2DProj JPetRecoImageTools::backProjectWithTOF(Matrix2DProj& sinogram, Matrix2DTOF& tof, int nAngles,
     RescaleFunc rescaleFunc, int rescaleMinCutoff, int rescaleFactor)
@@ -300,9 +275,23 @@ JPetRecoImageTools::Matrix2DProj JPetRecoImageTools::backProjectWithTOF(Matrix2D
         if (yMinusCenter2 + xMinusCenter2 < center2) {
           double t = ttemp - yMinusCenter * sin;
           int n = std::floor(t + 0.5F);
-          for (int i = 0; i < sinogram[n][angle]; i++) {
-            std::cout << normalDistributionProbability(std::sqrt(x * x + y * y), tof[n][angle][i], 100) << std::endl;
-            reconstructedProjection[y][x] += normalDistributionProbability(std::sqrt(x * x + y * y), tof[n][angle][i], 100);
+          auto tofInfo = tof.find(std::make_pair(n, angle));
+          std::vector<float> tofVector;
+          if (tofInfo != tof.end()) {
+            tofVector = tofInfo->second;
+          } else {
+            continue;
+          }
+          assert(sinogram[n][angle] == tofVector.size());
+          float lor_center_x = center + cos * (n - center);
+          float lor_center_y = center + sin * (n - center);
+
+          for (unsigned int i = 0; i < tofVector.size(); i++) {
+            float delta = std::abs(tofVector[i] * 0.299792458);
+            double distanceToCenterOfLOR = std::abs(std::sqrt(std::pow(lor_center_x - x, 2) + std::pow(lor_center_y - y, 2)));
+            double distributionProbability = normalDistributionProbability(distanceToCenterOfLOR, delta, 10);
+            distributionProbability = distributionProbability > 0.5 ? 1 - distributionProbability : distributionProbability;
+            reconstructedProjection[y][x] += distributionProbability;
           }
         }
       }
@@ -312,37 +301,29 @@ JPetRecoImageTools::Matrix2DProj JPetRecoImageTools::backProjectWithTOF(Matrix2D
   for (int x = 0; x < imageSize; x++) {
     for (int y = 0; y < imageSize; y++) {
       reconstructedProjection[y][x] *= angleStep;
-      reconstructedProjection[y][x] = std::round(reconstructedProjection[y][x]);
     }
   }
   rescaleFunc(reconstructedProjection, rescaleMinCutoff, rescaleFactor);
   return reconstructedProjection;
 }
 
-double normalDistributionProbability(float x, float mean, float stddev)
+double JPetRecoImageTools::normalDistributionProbability(float x, float mean, float stddev)
 {
-  double diff = x - mean;
-  return 1 / (stddev * std::sqrt(2 * M_PI)) * std::exp(-(diff * diff) / (2 * stddev * stddev));
+  return 0.5 * erfc(-(x - mean) / (stddev * sqrt(2)));
 }
 
-JPetRecoImageTools::Matrix2DProj JPetRecoImageTools::FilterSinogram(
-  JPetRecoImageTools::FourierTransformFunction& ftf,
-  JPetFilterInterface& filterFunction,
-  JPetRecoImageTools::Matrix2DProj& sinogram)
+JPetRecoImageTools::Matrix2DProj JPetRecoImageTools::FilterSinogram(JPetRecoImageTools::FourierTransformFunction& ftf,
+    JPetFilterInterface& filterFunction, JPetRecoImageTools::Matrix2DProj& sinogram)
 {
   return ftf(sinogram, filterFunction);
 }
 
 // see http://www.fftw.org/doc/One_002dDimensional-DFTs-of-Real-Data.html
 // http://www.fftw.org/fftw3.pdf
-
-JPetRecoImageTools::Matrix2DProj
-JPetRecoImageTools::doFFTW(Matrix2DProj& sinogram, JPetFilterInterface& filter)
+JPetRecoImageTools::Matrix2DProj JPetRecoImageTools::doFFTW(Matrix2DProj& sinogram, JPetFilterInterface& filter)
 {
-  >>> >>> > 496badb7b17333b4e657911055d72d41ad8b0125
   assert(sinogram.size() > 1);
-  JPetRecoImageTools::Matrix2DProj result(
-    sinogram.size(), std::vector< double >(sinogram[0].size()));
+  JPetRecoImageTools::Matrix2DProj result(sinogram.size(), std::vector<double>(sinogram[0].size()));
   int nAngles = sinogram[0].size();
   int nScanSize = sinogram.size();
   int inFTLength = std::floor(((nAngles / 2.)) + 1);
@@ -356,8 +337,7 @@ JPetRecoImageTools::doFFTW(Matrix2DProj& sinogram, JPetFilterInterface& filter)
   out = (fftw_complex*)fftw_malloc(outputSize * sizeof(fftw_complex));
   fftw_plan plan, invPlan;
   plan = fftw_plan_dft_r2c_2d(nScanSize, nAngles, in, out, FFTW_MEASURE);
-  invPlan =
-    fftw_plan_dft_c2r_2d(nScanSize, nAngles, out, outDouble, FFTW_MEASURE);
+  invPlan = fftw_plan_dft_c2r_2d(nScanSize, nAngles, out, outDouble, FFTW_MEASURE);
   for (int y = 0; y < nScanSize; y++) {
     for (int x = 0; x < nAngles; x++) {
       in[x + y * nAngles] = sinogram[y][x];
@@ -366,8 +346,7 @@ JPetRecoImageTools::doFFTW(Matrix2DProj& sinogram, JPetFilterInterface& filter)
   fftw_execute(plan);
 
   // max radius
-  double maxR = std::sqrt((inFTLength / 2. * inFTLength / 2.) +
-                          (nScanSize / 2. * nScanSize / 2.));
+  double maxR = std::sqrt((inFTLength / 2. * inFTLength / 2.) + (nScanSize / 2. * nScanSize / 2.));
 
   for (int x = 0; x < inFTLength; x++) {
     double xN = ((double)x - ((double)inFTLength / 2.));
@@ -399,19 +378,15 @@ JPetRecoImageTools::doFFTW(Matrix2DProj& sinogram, JPetFilterInterface& filter)
   return result;
 }
 
-
-JPetRecoImageTools::Matrix2DProj
-JPetRecoImageTools::doFFTSLOW(Matrix2DProj& sinogram,
-                              JPetFilterInterface& filter)
+JPetRecoImageTools::Matrix2DProj JPetRecoImageTools::doFFTSLOW(Matrix2DProj& sinogram, JPetFilterInterface& filter)
 {
   int nAngles = sinogram[0].size();
   int nScanSize = sinogram.size();
   int pow = std::round(std::log(nScanSize) / std::log(2.0));
   int padlen = std::round(std::pow(2.0, pow + 1));
-  JPetRecoImageTools::Matrix2DProj result(nScanSize,
-                                          std::vector< double >(nAngles));
-  std::vector< double > Re(padlen);
-  std::vector< double > Im(padlen);
+  JPetRecoImageTools::Matrix2DProj result(nScanSize, std::vector<double>(nAngles));
+  std::vector<double> Re(padlen);
+  std::vector<double> Im(padlen);
   for (int x = 0; x < nAngles; x++) {
     for (int y = 0; y < nScanSize; y++) {
       Re[y] = sinogram[y][x];
@@ -444,16 +419,13 @@ JPetRecoImageTools::doFFTSLOW(Matrix2DProj& sinogram,
   return result;
 }
 
-
-void JPetRecoImageTools::doFFTSLOWT(std::vector< double >& Re,
-                                    std::vector< double >& Im, int size,
-                                    int shift)
+void JPetRecoImageTools::doFFTSLOWT(std::vector<double>& Re, std::vector<double>& Im, int size, int shift)
 {
 
   int m = (int)(std::log((double)size) / std::log(2.0));
   int n = 1 << m;
-  std::vector< double > Imarg(n);
-  std::vector< double > Rearg(n);
+  std::vector<double> Imarg(n);
+  std::vector<double> Rearg(n);
 
   int i;
   for (i = 0; i < n; ++i) {
@@ -492,10 +464,8 @@ void JPetRecoImageTools::doFFTSLOWT(std::vector< double >& Re,
         int i_j = i + j;
         int i_j_s = i_j + stepsize;
         if (i > 0) {
-          Retmp =
-            Rearg[i << shifter] * Re[i_j_s] - Imarg[i << shifter] * Im[i_j_s];
-          Im[i_j_s] =
-            Rearg[i << shifter] * Im[i_j_s] + Imarg[i << shifter] * Re[i_j_s];
+          Retmp = Rearg[i << shifter] * Re[i_j_s] - Imarg[i << shifter] * Im[i_j_s];
+          Im[i_j_s] = Rearg[i << shifter] * Im[i_j_s] + Imarg[i << shifter] * Re[i_j_s];
           Re[i_j_s] = Retmp;
         }
 
@@ -512,10 +482,7 @@ void JPetRecoImageTools::doFFTSLOWT(std::vector< double >& Re,
   }
 }
 
-
-void JPetRecoImageTools::doFFTSLOWI(std::vector< double >& Re,
-                                    std::vector< double >& Im, int size,
-                                    int shift)
+void JPetRecoImageTools::doFFTSLOWI(std::vector<double>& Re, std::vector<double>& Im, int size, int shift)
 {
 
   for (int i = shift; i < shift + size; ++i) {
