+++ modules/large_barrel/JPetRecoImageTools/JPetRecoImageTools.h
@@ -18,6 +18,7 @@
 
 #include <cmath>
 #include <functional>
+#include <unordered_map>
 #include <vector>
 
 #include <cassert>
@@ -27,21 +28,22 @@
 
 #include "JPetFilterInterface.h"
 
-#include <boost/function_types/function_type.hpp>
-#include <boost/function_types/parameter_types.hpp>
-#include <boost/function_types/result_type.hpp>
+template <class T, typename U> struct PairHash {
+  size_t operator()(const std::pair<T, U>& key) const
+  {
+    return std::hash<T>()(key.first) ^ std::hash<U>()(key.second);
+  }
+};
 
 class JPetRecoImageTools
 {
 public:
-  using Matrix2D = std::vector< std::vector< int > >;
-  using Matrix2DProj = std::vector< std::vector< double > >;
-  using InterpolationFunc = std::function< double(
-      int i, double y, std::function< double(int, int) > &) >;
-  using RescaleFunc = std::function< void(Matrix2DProj &v, double minCutoff,
-                                          double rescaleFactor) >;
-  using FourierTransformFunction = std::function< Matrix2DProj(
-      Matrix2DProj &sinogram, JPetFilterInterface &filterFunction) >;
+  using Matrix2D = std::vector<std::vector<int>>;
+  using Matrix2DProj = std::vector<std::vector<double>>;
+  using Matrix2DTOF = std::unordered_map<std::pair<int, int>, std::vector<float>, PairHash<int, int>>;
+  using InterpolationFunc = std::function<double(int i, double y, std::function<double(int, int)>&)>;
+  using RescaleFunc = std::function<void(Matrix2DProj& v, double minCutoff, double rescaleFactor)>;
+  using FourierTransformFunction = std::function<Matrix2DProj(Matrix2DProj& sinogram, JPetFilterInterface& filterFunction)>;
 
   /// Returns a matrixGetter, that can be used to return matrix elements in the
   /// following way:
@@ -50,9 +52,7 @@ public:
   /// In addition if the indices goes outside of the matrix range 0 is retuned.
   /// It is assumed that the input matrix is quadratic.
   /// The produced functions can be used as an input to interpolation functions.
-  static std::function< double(int, int) >
-  matrixGetterFactory(const Matrix2D &emissionMatrix,
-                      bool isTransposed = false);
+  static std::function<double(int, int)> matrixGetterFactory(const Matrix2D& emissionMatrix, bool isTransposed = false);
 
   /*! \brief function returning func(i,j) where j is the nearest neighbour
    * index with respect to y.
@@ -60,27 +60,28 @@ public:
    *  \param y the double value for which the nearste neighoubring discrete
    * index is calculated.
    *  \param func function that returns double value based on two discrete i,j.
-  */
-  static double nearestNeighbour(int i, double y,
-                                 std::function< double(int, int) > &func);
+   */
+  static double nearestNeighbour(int i, double y, std::function<double(int, int)>& func);
   /*! \brief Linear interpolation function returning (1-t)*func(i,j) + t*
    * func(i,j+1).
    *  \param i discrete index being the first parameter of the function func.
    *  \param y the double value for which the j index  and t parameters are
    * calculated.
    *  \param func function that returns double value based on two discrete i,j.
-  */
-  static double linear(int i, double y,
-                       std::function< double(int, int) > &func);
+   */
+  static double linear(int i, double y, std::function<double(int, int)>& func);
 
   /// Rescale the Matrix in the following way:
   /// 1. All the values less than minCutoff are set to minCutoff
   /// 2. Removes the common backgroud term. So the values start at zero
   /// 3. Rescales all values by rescaleFactor/maxElement
   /// The final value range is from 0 to rescaleFactor
-  static void rescale(Matrix2DProj &v, double minCutoff, double rescaleFactor);
+  static void rescale(Matrix2DProj& v, double minCutoff, double rescaleFactor);
   /// PseudoRescale which does nothing
-  static void nonRescale(Matrix2DProj &, double, double) { return; }
+  static void nonRescale(Matrix2DProj&, double, double)
+  {
+    return;
+  }
 
   /*! \brief Function returning sinogram matrix.
    *  \param emissionMatrix matrix,  needs to be NxN
@@ -94,16 +95,12 @@ public:
    *  \param angleEnd end angle for projection in deg(Optional, default 180)
    *  \param rescaleFunc function that rescales the final result (Optional,
    * default no rescaling)
-  */
-  static Matrix2DProj createSinogramWithSingleInterpolation(
-      Matrix2D &emissionMatrix, int nViews, int nScans, double angleBeg = 0,
+   */
+  static Matrix2DProj createSinogramWithSingleInterpolation(Matrix2D& emissionMatrix, int nViews, int nScans, double angleBeg = 0,
       double angleEnd = 180, InterpolationFunc interpolationFunction = linear,
-      RescaleFunc rescaleFunc = nonRescale, int rescaleMinCutoff = 0,
-      int rescaleFactor = 255);
+      RescaleFunc rescaleFunc = nonRescale, int rescaleMinCutoff = 0, int rescaleFactor = 255);
 
-  static double calculateProjection(const Matrix2D &emissionMatrix, double phi,
-                                    int scanNumber, int nScans,
-                                    InterpolationFunc &interpolationFunction);
+  static double calculateProjection(const Matrix2D& emissionMatrix, double phi, int scanNumber, int nScans, InterpolationFunc& interpolationFunction);
 
   /*! \brief Function returning sinogram matrix with both variables interpolated
    *  \param emissionMatrix matrix, needs to be NxN
@@ -112,17 +109,12 @@ public:
    * default no rescaling)
    *  \param rescaleMinCutoff min value to set in rescale (Optional)
    *  \param rescaleFactor max value to set in rescale (Optional)
-  */
-  static Matrix2DProj
-  createSinogramWithDoubleInterpolation(Matrix2D &emissionMatrix, int nAngles,
-                                        RescaleFunc rescaleFunc = nonRescale,
-                                        int rescaleMinCutoff = 0,
-                                        int rescaleFactor = 255);
+   */
+  static Matrix2DProj createSinogramWithDoubleInterpolation(Matrix2D& emissionMatrix, int nAngles, RescaleFunc rescaleFunc = nonRescale,
+      int rescaleMinCutoff = 0, int rescaleFactor = 255);
 
-  static double
-  calculateProjection2(int step, double cos, double sin, int imageSize,
-                       double center, double length,
-                       std::function< double(int, int) > matrixGet);
+  static double calculateProjection2(int step, double cos, double sin, int imageSize, double center, double length,
+                                     std::function<double(int, int)> matrixGet);
 
   /*! \brief Function image from sinogram matrix
    *  \param sinogram matrix containing sinogram to backProject
@@ -131,10 +123,22 @@ public:
    * default no rescaling)
    *  \param rescaleMinCutoff min value to set in rescale (Optional)
    *  \param rescaleFactor max value to set in rescale (Optional)
-  */
-  static Matrix2DProj backProject(Matrix2DProj &sinogram, int angles,
-                                  RescaleFunc rescaleFunc, int rescaleMinCutoff,
-                                  int rescaleFactor);
+   */
+  static Matrix2DProj backProject(Matrix2DProj& sinogram, int angles, RescaleFunc rescaleFunc, int rescaleMinCutoff, int rescaleFactor);
+
+  /*! \brief Function image from sinogram matrix
+   *  \param sinogram matrix containing sinogram to backProject
+   *  \param tof vector with information about TOF for every hit
+   *  \param nAngles angle step is calculated as PI / nAngles
+   *  \param rescaleFunc function that rescales the final result (Optional,
+   * default no rescaling)
+   *  \param rescaleMinCutoff min value to set in rescale (Optional)
+   *  \param rescaleFactor max value to set in rescale (Optional)
+   */
+  static Matrix2DProj backProjectWithTOF(Matrix2DProj& sinogram, Matrix2DTOF& tof, int angles, RescaleFunc rescaleFunc, int rescaleMinCutoff,
+                                         int rescaleFactor);
+
+  static double normalDistributionProbability(float x, float mean, float stddev);
 
   /*! \brief Function filtering given sinogram using fouriner implementation and
    filter
@@ -142,35 +146,29 @@ public:
    *  \param filter type of filter
    *  \param sinogram data to filter
   */
-  static Matrix2DProj FilterSinogram(FourierTransformFunction &ftf,
-                                     JPetFilterInterface &filter,
-                                     Matrix2DProj &sinogram);
+  static Matrix2DProj FilterSinogram(FourierTransformFunction& ftf, JPetFilterInterface& filter, Matrix2DProj& sinogram);
 
   /*! \brief Fourier transform implementation using FFTW library
    *  \param sinogram data to filter
    *  \param filter type of filter
-  */
-  static Matrix2DProj doFFTW(Matrix2DProj &sinogram,
-                             JPetFilterInterface &filter);
+   */
+  static Matrix2DProj doFFTW(Matrix2DProj& sinogram, JPetFilterInterface& filter);
 
   /*! \brief Fourier transform implementation
    *  \param sinogram data to filter
    *  \param filter type of filter
    */
-  static Matrix2DProj doFFTSLOW(Matrix2DProj &sinogram,
-                                JPetFilterInterface &filter);
+  static Matrix2DProj doFFTSLOW(Matrix2DProj& sinogram, JPetFilterInterface& filter);
 
 private:
   JPetRecoImageTools();
   ~JPetRecoImageTools();
-  JPetRecoImageTools(const JPetRecoImageTools &) = delete;
-  JPetRecoImageTools &operator=(const JPetRecoImageTools &) = delete;
+  JPetRecoImageTools(const JPetRecoImageTools&) = delete;
+  JPetRecoImageTools& operator=(const JPetRecoImageTools&) = delete;
 
-  static void doFFTSLOWT(std::vector< double > &Re, std::vector< double > &Im,
-                         int size, int shift);
+  static void doFFTSLOWT(std::vector<double>& Re, std::vector<double>& Im, int size, int shift);
 
-  static void doFFTSLOWI(std::vector< double > &Re, std::vector< double > &Im,
-                         int size, int shift);
+  static void doFFTSLOWI(std::vector<double>& Re, std::vector<double>& Im, int size, int shift);
 
   static inline double setToZeroIfSmall(double value, double epsilon)
   {
