+++ modules/large_barrel/JPetRecoImageTools/JPetRecoImageToolsTest.cpp
@@ -1,1308 +1,498 @@
-#define BOOST_TEST_DYN_LINK
-#define BOOST_TEST_MODULE JPetRecoImageToolsTests
-#include <boost/test/unit_test.hpp>
-#include <fstream>
-#include <iostream>
-#include <utility>
-#include <vector>
+/**
+ *  @copyright Copyright 2016 The J-PET Framework Authors. All rights reserved.
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may find a copy of the License in the LICENCE file.
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ *  @file JPetRecoImageTools.cpp
+ */
 
-#include "JPetCommonTools/JPetCommonTools.h"
-#include "./JPetFilterCosine.h"
-#include "./JPetFilterHamming.h"
-#include "./JPetFilterNone.h"
-#include "./JPetFilterRamLak.h"
-#include "./JPetFilterRidgelet.h"
-#include "./JPetFilterSheppLogan.h"
 #include "./JPetRecoImageTools.h"
+#include "JPetLoggerInclude.h"
 
-BOOST_AUTO_TEST_SUITE(FirstSuite)
+JPetRecoImageTools::JPetRecoImageTools() {}
 
-BOOST_AUTO_TEST_CASE(rescale_0)
+JPetRecoImageTools::~JPetRecoImageTools() {}
+
+double JPetRecoImageTools::nearestNeighbour(int i, double y, std::function<double(int, int)>& func)
 {
-  JPetRecoImageTools::Matrix2DProj matrix = {{3, 0}, {0, 0}};
-  double resFact = 5;
-  double minCutoff = 0;
-  double epsilon = 0.00001;
-  JPetRecoImageTools::rescale(matrix, minCutoff, resFact);
-  BOOST_REQUIRE_CLOSE(matrix[0][0], 5, epsilon);
-  BOOST_REQUIRE_CLOSE(matrix[0][1], 0, epsilon);
-  BOOST_REQUIRE_CLOSE(matrix[1][0], 0, epsilon);
-  BOOST_REQUIRE_CLOSE(matrix[1][1], 0, epsilon);
+  int j = std::round(y);
+  return func(i, j);
 }
 
-BOOST_AUTO_TEST_CASE(rescale)
+double JPetRecoImageTools::linear(int i, double y, std::function<double(int, int)>& func)
 {
-  JPetRecoImageTools::Matrix2DProj matrix = {{2, 1}, {1, 1}};
-  double resFact = 2;
-  double minCutoff = 0;
-  double epsilon = 0.00001;
-  JPetRecoImageTools::rescale(matrix, minCutoff, resFact);
-  BOOST_REQUIRE_CLOSE(matrix[0][0], 2, epsilon);
-  BOOST_REQUIRE_CLOSE(matrix[0][1], 0, epsilon);
-  BOOST_REQUIRE_CLOSE(matrix[1][0], 0, epsilon);
-  BOOST_REQUIRE_CLOSE(matrix[1][1], 0, epsilon);
+  int j = std::floor(y);
+  double weight = std::abs(y - std::floor(y));
+  return (1 - weight) * func(i, j) + weight * func(i, j + 1);
 }
 
-BOOST_AUTO_TEST_CASE(rescale_2)
+std::function<double(int, int)> JPetRecoImageTools::matrixGetterFactory(const Matrix2D& emissionMatrix, bool isTransposed)
 {
-  JPetRecoImageTools::Matrix2DProj matrix = {{5, 2}, {1, 2}};
-  double resFact = 2;
-  double minCutoff = 0;
-  double epsilon = 0.00001;
-  JPetRecoImageTools::rescale(matrix, minCutoff, resFact);
-  BOOST_REQUIRE_CLOSE(matrix[0][0], 2, epsilon);
-  BOOST_REQUIRE_CLOSE(matrix[0][1], 0.5, epsilon);
-  BOOST_REQUIRE_CLOSE(matrix[1][0], 0., epsilon);
-  BOOST_REQUIRE_CLOSE(matrix[1][1], 0.5, epsilon);
+  if (!isTransposed) {
+    return [&emissionMatrix](int i, int j) {
+      if (i >= 0 && i < (int)emissionMatrix[0].size() && j >= 0 && j < (int)emissionMatrix.size()) {
+        return emissionMatrix[i][j];
+      } else {
+        return 0;
+      }
+    };
+  } else {
+    return [&emissionMatrix](int i, int j) {
+      if (i >= 0 && i < (int)emissionMatrix.size() && j >= 0 && j < (int)emissionMatrix[0].size()) {
+        return emissionMatrix[j][i];
+      } else {
+        return 0;
+      }
+    };
+  }
 }
 
-BOOST_AUTO_TEST_CASE(rescale_3)
+JPetRecoImageTools::Matrix2DProj JPetRecoImageTools::createSinogramWithSingleInterpolation(Matrix2D& emissionMatrix, int nViews, int nScans,
+    double angleBeg, double angleEnd,
+    InterpolationFunc interpolationFunction,
+    RescaleFunc rescaleFunc, int rescaleMinCutoff,
+    int rescaleFactor)
 {
-  JPetRecoImageTools::Matrix2DProj matrix = {{5, 3}, {1, 2}};
-  double resFact = 8;
-  double minCutoff = 2; /// -> 5, 3 , 2, 2
-  /// after background subtraction 3, 1, 0,0
-  double epsilon = 0.00001;
-  JPetRecoImageTools::rescale(matrix, minCutoff, resFact);
-  BOOST_REQUIRE_CLOSE(matrix[0][0], 8, epsilon);
-  BOOST_REQUIRE_CLOSE(matrix[0][1], 1. * 8. / 3., epsilon);
-  BOOST_REQUIRE_CLOSE(matrix[1][0], 0 * 8. / 3., epsilon);
-  BOOST_REQUIRE_CLOSE(matrix[1][1], 0, epsilon);
-}
-BOOST_AUTO_TEST_CASE(matrixGetterFactoryTest)
-{
-  double epsilon = 0.00001;
-  JPetRecoImageTools::Matrix2D matrix = {{1, 2}, {3, 4}};
-  auto getter = JPetRecoImageTools::matrixGetterFactory(matrix, false);
-  BOOST_REQUIRE_CLOSE(getter(0, 0), 1, epsilon);
-  BOOST_REQUIRE_CLOSE(getter(0, 1), 2, epsilon);
-  BOOST_REQUIRE_CLOSE(getter(1, 0), 3, epsilon);
-  BOOST_REQUIRE_CLOSE(getter(1, 1), 4, epsilon);
-  BOOST_REQUIRE_CLOSE(getter(3, 3), 0,
-                      epsilon); /// out of range should return 0
-  auto getterT =
-    JPetRecoImageTools::matrixGetterFactory(matrix, true); /// transposed
-  BOOST_REQUIRE_CLOSE(getterT(0, 0), 1, epsilon);
-  BOOST_REQUIRE_CLOSE(getterT(0, 1), 3, epsilon);
-  BOOST_REQUIRE_CLOSE(getterT(1, 0), 2, epsilon);
-  BOOST_REQUIRE_CLOSE(getterT(1, 1), 4, epsilon);
-  BOOST_REQUIRE_CLOSE(getterT(3, 3), 0,
-                      epsilon); /// out of range should return 0
+  assert(emissionMatrix.size() > 0);
+  assert(emissionMatrix.size() == emissionMatrix[0].size());
+  assert(nViews > 0);
+  assert(nScans > 0);
+  assert(angleBeg < angleEnd);
+  assert(rescaleMinCutoff < rescaleFactor);
+
+  Matrix2DProj proj(nScans, std::vector<double>(nViews));
+
+  float stepsize = (angleEnd - angleBeg) / nViews;
+  assert(stepsize > 0);
+
+  int viewIndex = 0;
+  for (auto phi = angleBeg; phi < angleEnd; phi = phi + stepsize, viewIndex++) {
+    for (auto scanNumber = 0; scanNumber < nScans; scanNumber++) {
+      proj[nScans - 1 - scanNumber][viewIndex] =
+        JPetRecoImageTools::calculateProjection(emissionMatrix, phi, scanNumber, nScans, interpolationFunction);
+    }
+  }
+  rescaleFunc(proj, rescaleMinCutoff, rescaleFactor);
+  return proj;
 }
 
-BOOST_AUTO_TEST_CASE(nearestNeighbour)
+double JPetRecoImageTools::calculateProjection(const Matrix2D& emissionMatrix, double angle, int scanNumber, int nScans,
+    InterpolationFunc& interpolationFunction)
 {
-  double epsilon = 0.00001;
-  JPetRecoImageTools::Matrix2D matrix = {{1, 2}, {3, 4}};
-  auto getter = JPetRecoImageTools::matrixGetterFactory(matrix, false);
-  /// first argument is i second j  (i,j)
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::nearestNeighbour(0, 0, getter), 1,
-                      epsilon);
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::nearestNeighbour(1, 0, getter), 3,
-                      epsilon);
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::nearestNeighbour(0, 0.6, getter), 2,
-                      epsilon); // (0,1)
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::nearestNeighbour(0, 0.4, getter), 1,
-                      epsilon); // (0,0)
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::nearestNeighbour(1, 0.4, getter), 3,
-                      epsilon); // (0,0)
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::nearestNeighbour(1, 0.8, getter), 4,
-                      epsilon); // (0,0)
-  auto getterT =
-    JPetRecoImageTools::matrixGetterFactory(matrix, true); /// transposed
-  /// first argument is j second i  (j,i)
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::nearestNeighbour(0, 0, getterT), 1,
-                      epsilon);
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::nearestNeighbour(1, 0, getterT), 2,
-                      epsilon);
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::nearestNeighbour(0, 0.6, getterT), 3,
-                      epsilon); // (1,0)
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::nearestNeighbour(0, 0.4, getterT), 1,
-                      epsilon); // (0,0)
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::nearestNeighbour(1, 0.4, getterT), 2,
-                      epsilon); // (0,1)
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::nearestNeighbour(1, 0.8, getterT), 4,
-                      epsilon); // (1,1)
+  int N = scanNumber - std::floor(nScans / 2);
+  const int kInputMatrixSize = emissionMatrix.size();
+  // if no. nScans is greater than the image width, then scale will be <1
+  const double scale = kInputMatrixSize / nScans;
+  const double kSin45or125deg = std::sqrt(2) / 2; /// sin(45) deg
+  const double kEpsilon = 0.0000001;
+  const double kDegToRad = M_PI / 180.;
+
+  double sin = std::sin(angle * kDegToRad - M_PI / 2.);
+  sin = setToZeroIfSmall(sin, kEpsilon);
+  double cos = std::cos(angle * kDegToRad - M_PI / 2.);
+  cos = setToZeroIfSmall(cos, kEpsilon);
+
+  double a = 0.;
+  double b = 0.;
+  /// The line over which we integrate is perpendicular to any line with the
+  /// slope = tg(angle), so it is always  -1/tg(angle).
+  /// If the angle is between (45 to 125)
+  /// we use  y = a * x + b, and we iterate over rows of the matrix (x).
+  /// If the angle is between [0 to 45 ] or [125 to 180]
+  /// we use x = a* y +b, and we iterate over columns of the matrix (y)
+  bool angleRange45To125 = std::abs(sin) > kSin45or125deg;
+  double divided = 1.;
+  std::function<double(int, int)> matrixGet;
+
+  if (angleRange45To125) {
+    assert(sin);
+    a = -cos / sin;
+    b = (N - cos - sin) / sin;
+    b *= scale;
+    matrixGet = matrixGetterFactory(emissionMatrix, false); // The matrix  elements will be taken as (x,y).
+    divided = std::abs(sin);
+  } else {
+    assert(cos);
+    a = -sin / cos;
+    b = (N - cos - sin) / cos;
+    b *= scale;
+    matrixGet = matrixGetterFactory(emissionMatrix, true); // The matrix  elements will be taken as (y, x) - transposed.
+    divided = std::abs(cos);
+  }
+  const int kMatrixCenter = std::floor(emissionMatrix.size() / 2);
+  double value = 0.;
+  for (auto i = -kMatrixCenter; i < kMatrixCenter; i++) {
+    value += interpolationFunction(i + kMatrixCenter, a * i + b + kMatrixCenter, matrixGet);
+  }
+  value /= divided;
+  return value;
 }
 
-BOOST_AUTO_TEST_CASE(linear)
+JPetRecoImageTools::Matrix2DProj JPetRecoImageTools::createSinogramWithDoubleInterpolation(Matrix2D& emissionMatrix, int nAngles,
+    RescaleFunc rescaleFunc, int rescaleMinCutoff,
+    int rescaleFactor)
 {
-  double epsilon = 0.00001;
-  JPetRecoImageTools::Matrix2D matrix = {{1, 2}, {3, 4}};
-  auto getter = JPetRecoImageTools::matrixGetterFactory(matrix, false);
-  /// first argument is i second j  (i,j)
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::linear(0, 0, getter), 1, epsilon);
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::linear(0, 1, getter), 2, epsilon);
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::linear(0, 0.9, getter), 1.9, epsilon);
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::linear(1, 0, getter), 3, epsilon);
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::linear(0, 0.4, getter), 1.4,
-                      epsilon); // 1 * 0.6 + 2 * 0.4
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::linear(1, 0.4, getter), 3.4,
-                      epsilon); //  3* 0.6 + 4 * 0.4
-  auto getterT =
-    JPetRecoImageTools::matrixGetterFactory(matrix, true); /// transposed
-  /// first argument is j second i  (j,i)
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::linear(0, 0, getterT), 1, epsilon);
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::linear(1, 0, getterT), 2, epsilon);
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::linear(0, 0.6, getterT), 2.2,
-                      epsilon); // 1*0.4 +  3  *0.6
-}
-/*
-BOOST_AUTO_TEST_CASE(calculateProjection)
-{
-  double epsilon = 0.00001;
-  JPetRecoImageTools::Matrix2D matrix = {{4}};
-  //static double calculateProjection(const Matrix2D& emissionMatrix,
-  //double phi,
-  //int scanNumber,
-  //int nScans,
-  //InterpolationFunc& interpolationFunction
-  //);
-  /// first argument is i second j  (i,j)
-  JPetRecoImageTools::InterpolationFunc func =
-JPetRecoImageTools::nearestNeighbour;
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,0,0,1,
-func), 4, epsilon);
+  assert(emissionMatrix.size() > 0);
+  assert(emissionMatrix.size() == emissionMatrix[0].size());
+  assert(nAngles > 0);
+  assert(rescaleMinCutoff < rescaleFactor);
+  int imageSize = emissionMatrix[0].size();
+  double center = (double)(imageSize - 1) / 2.0;
+  double center2 = center * center;
+  double angleStep = M_PI / nAngles;
+
+  Matrix2DProj proj(imageSize, std::vector<double>(nAngles));
+  std::function<double(int, int)> matrixGet;
+  matrixGet = matrixGetterFactory(emissionMatrix, false); // The matrix  elements will be taken as (x,y).
+  for (int angle = 0; angle < nAngles; angle++) {
+    double cos = std::cos((double)angle * angleStep);
+    double sin = std::sin((double)angle * angleStep);
+    for (int scanNumber = 0; scanNumber < imageSize - 1; scanNumber++) {
+      proj[scanNumber][angle] = calculateProjection2(scanNumber, cos, sin, imageSize, center, center2, matrixGet);
+    }
+  }
+  rescaleFunc(proj, rescaleMinCutoff, rescaleFactor);
+  return proj;
 }
 
-BOOST_AUTO_TEST_CASE(calculateProjection2)
+double JPetRecoImageTools::calculateProjection2(int step, double cos, double sin, int imageSize, double center, double center2,
+    std::function<double(int, int)> matrixGet)
 {
-double epsilon = 0.00001;
-JPetRecoImageTools::Matrix2D matrix = {{1, 2}, {3, 4}};
-/// 1 2
-/// 3 4
-/// angle = 0
-/// nScans = 2
-JPetRecoImageTools::InterpolationFunc func =
-JPetRecoImageTools::nearestNeighbour;
-const int kNScans = 2;
-auto angle = 0.0;
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,angle,0,kNScans,
-func), 6, epsilon); // 2 + 4
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,angle,1,kNScans,
-func), 4, epsilon); // 1 + 3
-//BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,45,0,kNScans,
-func), 5, epsilon); // 3 + 2
-//BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,45,1,kNScans,
-func), 4, epsilon); // 4
-//BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,45,-1,kNScans,
-func), 1, epsilon); // 1
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,90,0,kNScans,
-func), 7, epsilon); // 3 + 4
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,90,1,kNScans,
-func), 3, epsilon); // 1 + 2
-//BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,125,0,kNScans,
-func), 5, epsilon); // 1 + 4
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,180,0,kNScans,
-func), 6, epsilon); // 2 + 4
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,180,1,kNScans,
-func), 4, epsilon); // 1 + 3
+  double stepMinusCenter = step - center;
+  double xtmp = center + stepMinusCenter * cos;
+  double ytmp = center - stepMinusCenter * sin;
+  int nmin = step == 0 || step == imageSize - 1 ? 0 : std::floor(center - std::sqrt((center2 - (stepMinusCenter * stepMinusCenter))));
+  double p = 0.0;
+  for (int n = nmin; n < imageSize - nmin; n++) {
+    double nMinusCenter = (double)n - center;
+    int x = std::floor((xtmp - nMinusCenter * sin) + 0.5);
+    int y = std::floor((ytmp - nMinusCenter * cos) + 0.5);
+    p += matrixGet(y, x); // y - height, x - widht
+  }
+  return p;
 }
 
-BOOST_AUTO_TEST_CASE(calculateProjection3)
+void JPetRecoImageTools::rescale(Matrix2DProj& matrix, double minCutoff, double rescaleFactor)
 {
-double epsilon = 0.00001;
-JPetRecoImageTools::Matrix2D matrix = {{1, 2, 3}, {4, 5, 6}, {7,8,9}};
-/// 1 2 3
-/// 4 5 6
-/// 7 8 9
-/// angle = 0
-/// nScans = 3
-JPetRecoImageTools::InterpolationFunc func =
-JPetRecoImageTools::nearestNeighbour;
-const int kNScans = 3;
-auto angle = 0.0;
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,angle,0,kNScans,
-func), 12, epsilon);
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,angle,1,kNScans,
-func), 15, epsilon);
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,angle,2,kNScans,
-func), 18, epsilon);
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,45,0,kNScans,
-func), 15, epsilon);
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,90,0,kNScans,
-func), 6, epsilon);
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,90,1,kNScans,
-func), 15, epsilon);
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,90,2,kNScans,
-func), 24, epsilon);
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,125,0,kNScans,
-func), 15, epsilon);
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,180,0,kNScans,
-func), 12, epsilon);
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,180,1,kNScans,
-func), 15, epsilon);
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,180,2,kNScans,
-func), 18, epsilon);
+
+  double datamax = matrix[0][0];
+  double datamin = matrix[0][0];
+  for (unsigned int k = 0; k < matrix.size(); k++) {
+    for (unsigned int j = 0; j < matrix[k].size(); j++) {
+      /// Applying min Cutoff
+      if (matrix[k][j] < minCutoff)
+        matrix[k][j] = minCutoff;
+      /// Finding the largest and the smallest element in the matrix.
+      if (matrix[k][j] > datamax)
+        datamax = matrix[k][j];
+      if (matrix[k][j] < datamin)
+        datamin = matrix[k][j];
+    }
+  }
+
+  /// datamin represents the constant background factor.
+  if ((datamax - datamin) == 0.) {
+    WARNING("Maximum value in the matrix to rescale is 0. No rescaling performed.");
+    return;
+  }
+
+  for (unsigned int k = 0; k < matrix.size(); k++) {
+    for (unsigned int j = 0; j < matrix[0].size(); j++) {
+      matrix[k][j] = (matrix[k][j] - datamin) * rescaleFactor / (datamax - datamin);
+    }
+  }
 }
 
-BOOST_AUTO_TEST_CASE(calculateProjection4)
+JPetRecoImageTools::Matrix2DProj JPetRecoImageTools::backProject(Matrix2DProj& sinogram, int nAngles, RescaleFunc rescaleFunc, int rescaleMinCutoff,
+    int rescaleFactor)
 {
-double epsilon = 0.00001;
-JPetRecoImageTools::Matrix2D matrix = {{1, 2, 3, 4}, {5, 6, 7, 8}, {9, 10, 11,
-12}, {13, 14, 15, 16}};
-/// 1  2  3  4
-/// 5  6  7  8
-/// 9  10 11 12
-/// 13 14 15 16
-/// angle = 0
-/// nScans = 4
-JPetRecoImageTools::InterpolationFunc func =
-JPetRecoImageTools::nearestNeighbour;
-const int kNScans = 4;
-auto angle = 0.0;
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,angle,0,kNScans,
-func), 28, epsilon);
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,angle,1,kNScans,
-func), 32, epsilon);
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,angle,2,kNScans,
-func), 36, epsilon);
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,angle,3,kNScans,
-func), 40, epsilon);
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,45,0,kNScans,
-func), 34, epsilon);
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,90,0,kNScans,
-func), 10, epsilon);
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,90,1,kNScans,
-func), 26, epsilon);
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,90,2,kNScans,
-func), 42, epsilon);
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,90,3,kNScans,
-func), 58, epsilon);
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,125,0,kNScans,
-func), 34, epsilon);
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,180,0,kNScans,
-func), 28, epsilon);
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,180,1,kNScans,
-func), 32, epsilon);
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,180,2,kNScans,
-func), 36, epsilon);
-BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection(matrix,180,3,kNScans,
-func), 40, epsilon);
+  int imageSize = sinogram.size();
+  double center = (double)(imageSize - 1) / 2.0;
+  double center2 = center * center;
+  double angleStep = M_PI / (double)nAngles;
+
+  Matrix2DProj reconstructedProjection(imageSize, std::vector<double>(imageSize));
+
+  for (int angle = 0; angle < nAngles; angle++) {
+    double cos = std::cos((double)angle * (double)angleStep);
+    double sin = std::sin((double)angle * (double)angleStep);
+
+    for (int x = 0; x < imageSize; x++) {
+      double xMinusCenter = (double)x - center;
+      double xMinusCenter2 = xMinusCenter * xMinusCenter;
+      double ttemp = xMinusCenter * cos + center;
+
+      for (int y = 0; y < imageSize; y++) {
+        double yMinusCenter = (double)y - center;
+        double yMinusCenter2 = yMinusCenter * yMinusCenter;
+        if (yMinusCenter2 + xMinusCenter2 < center2) {
+          double t = ttemp - yMinusCenter * sin;
+          int n = std::floor(t + 0.5F);
+          reconstructedProjection[y][x] += sinogram[n][angle];
+        }
+      }
+    }
+  }
+
+  for (int x = 0; x < imageSize; x++) {
+    for (int y = 0; y < imageSize; y++) {
+      reconstructedProjection[y][x] *= angleStep;
+    }
+  }
+  rescaleFunc(reconstructedProjection, rescaleMinCutoff, rescaleFactor);
+  return reconstructedProjection;
 }
 
-BOOST_AUTO_TEST_CASE(sinogram1)
+JPetRecoImageTools::Matrix2DProj JPetRecoImageTools::backProjectWithTOF(Matrix2DProj& sinogram, Matrix2DTOF& tof, int nAngles,
+    RescaleFunc rescaleFunc, int rescaleMinCutoff, int rescaleFactor)
 {
-double epsilon = 0.00001;
-JPetRecoImageTools::Matrix2D matrix = {{1, 2, 3}, {4, 5, 6}, {7,8,9}};
-/// 1 2 3
-/// 4 5 6
-/// 7 8 9
-JPetRecoImageTools::InterpolationFunc func =
-JPetRecoImageTools::nearestNeighbour;
-JPetRecoImageTools::RescaleFunc func2= JPetRecoImageTools::nonRescale;
-const int kNScans = 3;
-const int kNViews = 2;
-auto results = JPetRecoImageTools::sinogram(matrix,kNViews,kNScans, 0 , 180,
-func, func2);
-for (auto i: results)
-for (auto j: i)
-std::cout <<j <<std::endl;
+  int imageSize = sinogram.size();
+  double center = (double)(imageSize - 1) / 2.0;
+  double center2 = center * center;
+  double angleStep = M_PI / (double)nAngles;
+
+  Matrix2DProj reconstructedProjection(imageSize, std::vector<double>(imageSize));
+
+  for (int angle = 0; angle < nAngles; angle++) {
+    double cos = std::cos((double)angle * (double)angleStep);
+    double sin = std::sin((double)angle * (double)angleStep);
+
+    for (int x = 0; x < imageSize; x++) {
+      double xMinusCenter = (double)x - center;
+      double xMinusCenter2 = xMinusCenter * xMinusCenter;
+      double ttemp = xMinusCenter * cos + center;
+
+      for (int y = 0; y < imageSize; y++) {
+        double yMinusCenter = (double)y - center;
+        double yMinusCenter2 = yMinusCenter * yMinusCenter;
+        if (yMinusCenter2 + xMinusCenter2 < center2) {
+          double t = ttemp - yMinusCenter * sin;
+          int n = std::floor(t + 0.5F);
+          auto tofInfo = tof.find(std::make_pair(n, angle));
+          std::vector<float> tofVector;
+          if (tofInfo != tof.end()) {
+            tofVector = tofInfo->second;
+          } else {
+            continue;
+          }
+          assert(sinogram[n][angle] == tofVector.size());
+          float lor_center_x = center + cos * (n - center);
+          float lor_center_y = center + sin * (n - center);
+
+          for (unsigned int i = 0; i < tofVector.size(); i++) {
+            float delta = std::abs(tofVector[i] * 0.299792458);
+            double distanceToCenterOfLOR = std::abs(std::sqrt(std::pow(lor_center_x - x, 2) + std::pow(lor_center_y - y, 2)));
+            double distributionProbability = normalDistributionProbability(distanceToCenterOfLOR, delta, 10);
+            distributionProbability = distributionProbability > 0.5 ? 1 - distributionProbability : distributionProbability;
+            reconstructedProjection[y][x] += distributionProbability;
+          }
+        }
+      }
+    }
+  }
+
+  for (int x = 0; x < imageSize; x++) {
+    for (int y = 0; y < imageSize; y++) {
+      reconstructedProjection[y][x] *= angleStep;
+    }
+  }
+  rescaleFunc(reconstructedProjection, rescaleMinCutoff, rescaleFactor);
+  return reconstructedProjection;
 }
 
-//BOOST_AUTO_TEST_CASE(sinogram2)
-//{
-//double epsilon = 0.00001;
-//JPetRecoImageTools::Matrix2D matrix = {{0, 0, 0}, {0, 10, 0}, {0,0,0}};
-//JPetRecoImageTools::InterpolationFunc func =
-JPetRecoImageTools::nearestNeighbour;
-//JPetRecoImageTools::RescaleFunc func2= JPetRecoImageTools::nonRescale;
-//const int kNScans = 3;
-//const int kNViews = 4;
-//auto results = JPetRecoImageTools::sinogram(matrix,kNViews,kNScans, 0 , 180,
-func, func2);
-//for (auto i: results)
-//for (auto j: i)
-//std::cout <<j <<std::endl;
-//}
-*/
-BOOST_AUTO_TEST_CASE(projection1)
+double JPetRecoImageTools::normalDistributionProbability(float x, float mean, float stddev)
 {
-  double epsilon = 0.00001;
-  JPetRecoImageTools::Matrix2D matrix = {{1, 2, 3}, {4, 5, 6}, {7, 8, 9}};
-  /// 1 2 3
-  /// 4 5 6
-  /// 7 8 9
-
-  int imageSize = matrix.size();
-  double center = (float)(imageSize - 1) / 2.0;
-  double length = center * center;
-  std::function< double(int, int) > matrixGet;
-  matrixGet = JPetRecoImageTools::matrixGetterFactory(
-                matrix, false); // The matrix  elements will be taken as (x,y).
-  double step = (M_PI / 180);
-
-  double cos = std::cos(step * 0); // 0 degree
-  double sin = std::sin(step * 0); // 0 degree
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        0, cos, sin, imageSize, center, length, matrixGet),
-                      12, epsilon); // 1 + 4 + 7
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        1, cos, sin, imageSize, center, length, matrixGet),
-                      15, epsilon); // 2 + 5 + 8
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        2, cos, sin, imageSize, center, length, matrixGet),
-                      18, epsilon); // 3 + 6 + 9
-
-  cos = std::cos(step * 90); // 90 degree
-  sin = std::sin(step * 90); // 90 degree
-
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        0, cos, sin, imageSize, center, length, matrixGet),
-                      24, epsilon); // 9 + 8 + 7
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        1, cos, sin, imageSize, center, length, matrixGet),
-                      15, epsilon); // 6 + 5 + 3
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        2, cos, sin, imageSize, center, length, matrixGet),
-                      6, epsilon); // 3 + 2 + 1
+  return 0.5 * erfc(-(x - mean) / (stddev * sqrt(2)));
 }
 
-BOOST_AUTO_TEST_CASE(projection2)
+JPetRecoImageTools::Matrix2DProj JPetRecoImageTools::FilterSinogram(JPetRecoImageTools::FourierTransformFunction& ftf,
+    JPetFilterInterface& filterFunction, JPetRecoImageTools::Matrix2DProj& sinogram)
 {
-  double epsilon = 0.00001;
-  JPetRecoImageTools::Matrix2D matrix = {{1, 2}, {3, 4}};
-  /// 1 2
-  /// 3 4
-
-  int imageSize = matrix.size();
-  double center = (float)(imageSize - 1) / 2.0;
-  double length = center * center;
-  std::function< double(int, int) > matrixGet;
-  matrixGet = JPetRecoImageTools::matrixGetterFactory(
-                matrix, false); // The matrix  elements will be taken as (x,y).
-  double step = (M_PI / 180);
-
-  double cos = std::cos(step * 0); // 0 degree
-  double sin = std::sin(step * 0); // 0 degree
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        0, cos, sin, imageSize, center, length, matrixGet),
-                      4, epsilon); // 1 + 3
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        1, cos, sin, imageSize, center, length, matrixGet),
-                      6, epsilon); // 2 + 4
-
-  cos = std::cos(step * 90); // 90 degree
-  sin = std::sin(step * 90); // 90 degree
-
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        0, cos, sin, imageSize, center, length, matrixGet),
-                      7, epsilon); // 3 + 4
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        1, cos, sin, imageSize, center, length, matrixGet),
-                      3, epsilon); // 1 + 2
+  return ftf(sinogram, filterFunction);
 }
 
-BOOST_AUTO_TEST_CASE(projection3)
+// see http://www.fftw.org/doc/One_002dDimensional-DFTs-of-Real-Data.html
+// http://www.fftw.org/fftw3.pdf
+JPetRecoImageTools::Matrix2DProj JPetRecoImageTools::doFFTW(Matrix2DProj& sinogram, JPetFilterInterface& filter)
 {
-  double epsilon = 0.00001;
-  JPetRecoImageTools::Matrix2D matrix = {{1, 2, 3, 4, 5},
-    {6, 7, 8, 9, 10},
-    {11, 12, 13, 14, 15},
-    {16, 17, 18, 19, 20},
-    {21, 22, 23, 24, 25}
-  };
-  /// 1  2  3  4  5
-  /// 6  7  8  9  10
-  /// 11 12 13 14 15
-  /// 16 17 18 19 20
-  /// 21 22 23 24 25
-  /// angle = 0
-  /// nScans = 4
-  int imageSize = matrix.size();
-  double center = (float)(imageSize - 1) / 2.0;
-  double length = center * center;
-  std::function< double(int, int) > matrixGet;
-  matrixGet = JPetRecoImageTools::matrixGetterFactory(
-                matrix, false); // The matrix  elements will be taken as (x,y).
-  double step = (M_PI / 180);
+  assert(sinogram.size() > 1);
+  JPetRecoImageTools::Matrix2DProj result(sinogram.size(), std::vector<double>(sinogram[0].size()));
+  int nAngles = sinogram[0].size();
+  int nScanSize = sinogram.size();
+  int inFTLength = std::floor(((nAngles / 2.)) + 1);
+  int outputSize = nScanSize * inFTLength;
+  int size = nScanSize * nAngles;
+  double* in;
+  in = (double*)malloc(size * sizeof(double));
+  double* outDouble;
+  outDouble = (double*)malloc(size * sizeof(double));
+  fftw_complex* out;
+  out = (fftw_complex*)fftw_malloc(outputSize * sizeof(fftw_complex));
+  fftw_plan plan, invPlan;
+  plan = fftw_plan_dft_r2c_2d(nScanSize, nAngles, in, out, FFTW_MEASURE);
+  invPlan = fftw_plan_dft_c2r_2d(nScanSize, nAngles, out, outDouble, FFTW_MEASURE);
+  for (int y = 0; y < nScanSize; y++) {
+    for (int x = 0; x < nAngles; x++) {
+      in[x + y * nAngles] = sinogram[y][x];
+    }
+  }
+  fftw_execute(plan);
 
-  double cos = std::cos(step * 0); // 0 degree
-  double sin = std::sin(step * 0); // 0 degree
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        0, cos, sin, imageSize, center, length, matrixGet),
-                      55, epsilon); // 21 + 16 + 11 + 6 + 1
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        1, cos, sin, imageSize, center, length, matrixGet),
-                      60, epsilon); // 22 + 17 + 12 + 7 + 2
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        2, cos, sin, imageSize, center, length, matrixGet),
-                      65, epsilon); // 23 + 18 + 13 + 8 + 3
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        3, cos, sin, imageSize, center, length, matrixGet),
-                      70, epsilon); // 24 + 19 + 14 + 9 + 4
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        4, cos, sin, imageSize, center, length, matrixGet),
-                      75, epsilon); // 25 + 20 + 15 + 10 + 5
+  // max radius
+  double maxR = std::sqrt((inFTLength / 2. * inFTLength / 2.) + (nScanSize / 2. * nScanSize / 2.));
 
-  // cos = std::cos(step * 45); //45 degree
-  // sin = std::sin(step * 45); //45 degree
-  //
-  // BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(0,
-  //               cos, sin, imageSize, center, length, matrixGet), 70,
-  //               epsilon); // 24 + 19 + 14 + 9 + 4
+  for (int x = 0; x < inFTLength; x++) {
+    double xN = ((double)x - ((double)inFTLength / 2.));
+    for (int y = 0; y < nScanSize; y++) {
+      // go by columns
+      uint gid = y * inFTLength + x;
+      // shifting coordinates
+      double yN = ((double)y - ((double)nScanSize / 2.));
 
-  cos = std::cos(step * 90); // 90 degree
-  sin = std::sin(step * 90); // 90 degree
+      // current radius normalized to [0 .. 1]
+      double r = std::sqrt((xN * xN + yN * yN)) / maxR;
+      double filterValue = std::abs(filter(1 - r));
+      out[gid][0] *= filterValue;
+      out[gid][1] *= filterValue;
+    }
+  }
+  fftw_execute(invPlan);
 
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        0, cos, sin, imageSize, center, length, matrixGet),
-                      115, epsilon); // 25 + 24 + 23 + 22 + 21
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        1, cos, sin, imageSize, center, length, matrixGet),
-                      90, epsilon); // 20 + 19 + 18 + 17 + 16
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        2, cos, sin, imageSize, center, length, matrixGet),
-                      65, epsilon); // 15 + 14 + 13 + 12 + 11
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        3, cos, sin, imageSize, center, length, matrixGet),
-                      40, epsilon); // 10 + 9 + 8 + 7 + 6
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        4, cos, sin, imageSize, center, length, matrixGet),
-                      15, epsilon); // 5 + 4 + 3 + 2 + 1
+  for (int y = 0; y < nScanSize; y++) {
+    for (int x = 0; x < nAngles; x++) {
+      result[y][x] = outDouble[x + y * nAngles] / size;
+    }
+  }
+
+  fftw_free(out);
+  fftw_destroy_plan(plan);
+  fftw_destroy_plan(invPlan);
+  fftw_cleanup();
+  return result;
 }
 
-BOOST_AUTO_TEST_CASE(projection4)
+JPetRecoImageTools::Matrix2DProj JPetRecoImageTools::doFFTSLOW(Matrix2DProj& sinogram, JPetFilterInterface& filter)
 {
-  double epsilon = 0.00001;
-  JPetRecoImageTools::Matrix2D matrix = {
-    {1, 10, 100, 1000, 10000, 100000, 1000000},
-    {2, 20, 200, 2000, 20000, 200000, 2000000},
-    {3, 30, 300, 3000, 30000, 300000, 3000000},
-    {4, 40, 400, 4000, 40000, 400000, 4000000},
-    {5, 50, 500, 5000, 50000, 500000, 5000000},
-    {6, 60, 600, 6000, 60000, 600000, 6000000},
-    {7, 70, 700, 7000, 70000, 700000, 7000000}
-  };
-  int imageSize = matrix.size();
-  double center = (float)(imageSize - 1) / 2.0;
-  double length = center * center;
-  std::function< double(int, int) > matrixGet;
-  matrixGet = JPetRecoImageTools::matrixGetterFactory(
-                matrix, false); // The matrix  elements will be taken as (x,y).
-  double step = (M_PI / 180);
+  int nAngles = sinogram[0].size();
+  int nScanSize = sinogram.size();
+  int pow = std::round(std::log(nScanSize) / std::log(2.0));
+  int padlen = std::round(std::pow(2.0, pow + 1));
+  JPetRecoImageTools::Matrix2DProj result(nScanSize, std::vector<double>(nAngles));
+  std::vector<double> Re(padlen);
+  std::vector<double> Im(padlen);
+  for (int x = 0; x < nAngles; x++) {
+    for (int y = 0; y < nScanSize; y++) {
+      Re[y] = sinogram[y][x];
+      Im[y] = 0.;
+    }
+    for (int y = nScanSize; y < padlen; y++) {
+      Re[y] = 0.;
+      Im[y] = 0.;
+    }
+    doFFTSLOWT(Re, Im, padlen, 0);
+    for (int y = 0; y < (padlen / 2); y++) {
+      double highest = padlen / 2; // highest value is on middle of 1D column
+      double r = y / highest;      // distance from highest value
+      double filterValue = std::abs(filter(r));
 
-  double cos = std::cos(step * 0); // 0 degree
-  double sin = std::sin(step * 0); // 0 degree
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        0, cos, sin, imageSize, center, length, matrixGet),
-                      28, epsilon);
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        1, cos, sin, imageSize, center, length, matrixGet),
-                      280, epsilon);
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        2, cos, sin, imageSize, center, length, matrixGet),
-                      2800, epsilon);
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        3, cos, sin, imageSize, center, length, matrixGet),
-                      28000, epsilon);
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        4, cos, sin, imageSize, center, length, matrixGet),
-                      280000, epsilon);
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        5, cos, sin, imageSize, center, length, matrixGet),
-                      2800000, epsilon);
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        6, cos, sin, imageSize, center, length, matrixGet),
-                      28000000, epsilon);
+      Re[y] *= filterValue;
+      Im[y] *= filterValue;
+      Re[padlen - y] *= filterValue;
+      Im[padlen - y] *= filterValue;
+    }
 
-  cos = std::cos(step * 90); // 90 degree
-  sin = std::sin(step * 90); // 90 degree
+    Re[padlen / 2] *= std::abs(filter(1));
+    Im[padlen / 2] *= std::abs(filter(1));
 
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        0, cos, sin, imageSize, center, length, matrixGet),
-                      7777777, epsilon);
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        1, cos, sin, imageSize, center, length, matrixGet),
-                      6666666, epsilon);
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        2, cos, sin, imageSize, center, length, matrixGet),
-                      5555555, epsilon);
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        3, cos, sin, imageSize, center, length, matrixGet),
-                      4444444, epsilon);
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        4, cos, sin, imageSize, center, length, matrixGet),
-                      3333333, epsilon);
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        5, cos, sin, imageSize, center, length, matrixGet),
-                      2222222, epsilon);
-  BOOST_REQUIRE_CLOSE(JPetRecoImageTools::calculateProjection2(
-                        6, cos, sin, imageSize, center, length, matrixGet),
-                      1111111, epsilon);
+    doFFTSLOWI(Re, Im, padlen, 0);
+    for (int y = 0; y < nScanSize; y++) {
+      result[y][x] = Re[y];
+    }
+  }
+  return result;
 }
 
-// This test takes a Shepp-Logan phantom and creates a sinogram.
-BOOST_AUTO_TEST_CASE(sinogram)
+void JPetRecoImageTools::doFFTSLOWT(std::vector<double>& Re, std::vector<double>& Im, int size, int shift)
 {
-  const auto inFile = "unitTestData/JPetRecoImageToolsTest/phantom.pgm";
-  const auto outFile = "sinogram.ppm";
-  /// read phantom
-  std::ifstream in(inFile);
-  BOOST_REQUIRE(in);
-  std::string line;
-  getline(in, line);
-  unsigned int width;
-  unsigned int height;
-  in >> width;
-  in >> height;
-  int val;
-  in >> val; // skip max val
-  std::vector< std::vector< int > > m(width, std::vector< int >(height));
-  for (unsigned int i = 0; i < width; i++) {
-    for (unsigned int j = 0; j < height; j++) {
-      in >> val;
-      m[i][j] = val;
-    }
+
+  int m = (int)(std::log((double)size) / std::log(2.0));
+  int n = 1 << m;
+  std::vector<double> Imarg(n);
+  std::vector<double> Rearg(n);
+
+  int i;
+  for (i = 0; i < n; ++i) {
+    double arg = 2 * M_PI * (double)((float)i) / (double)((float)n);
+    Rearg[i] = std::cos(arg);
+    Imarg[i] = -std::sin(arg);
   }
-  int views = 180;
-  int scans = 256;
-  std::vector< std::vector< double > > result =
-    JPetRecoImageTools::createSinogramWithSingleInterpolation(
-      m, views, scans, 0, 180, JPetRecoImageTools::linear,
-      JPetRecoImageTools::rescale);
-  /// save sinogram
-  std::ofstream res(outFile);
-  res << "P2" << std::endl;
-  res << result[0].size() << " " << result.size() << std::endl;
-  res << "255" << std::endl;
-  for (unsigned int i = 0; i < result.size(); i++) {
-    for (unsigned int j = 0; j < result[0].size(); j++) {
-      res << static_cast< int >(result[i][j]) << " ";
+
+  int j = shift;
+
+  double Retmp;
+  double Imtmp;
+  for (i = shift; i < shift + n - 1; ++i) {
+    if (i < j) {
+      Retmp = Re[i];
+      Imtmp = Im[i];
+      Re[i] = Re[j];
+      Im[i] = Im[j];
+      Re[j] = Retmp;
+      Im[j] = Imtmp;
     }
-    res << std::endl;
+
+    int k;
+    for (k = n >> 1; k + shift <= j; k /= 2) {
+      j -= k;
+    }
+
+    j += k;
   }
-  res.close();
-  BOOST_REQUIRE(JPetCommonTools::ifFileExisting(outFile));
+
+  int stepsize = 1;
+
+  for (int shifter = m - 1; stepsize < n; --shifter) {
+    for (j = shift; j < shift + n; j += stepsize << 1) {
+      for (i = 0; i < stepsize; ++i) {
+        int i_j = i + j;
+        int i_j_s = i_j + stepsize;
+        if (i > 0) {
+          Retmp = Rearg[i << shifter] * Re[i_j_s] - Imarg[i << shifter] * Im[i_j_s];
+          Im[i_j_s] = Rearg[i << shifter] * Im[i_j_s] + Imarg[i << shifter] * Re[i_j_s];
+          Re[i_j_s] = Retmp;
+        }
+
+        Retmp = Re[i_j] - Re[i_j_s];
+        Imtmp = Im[i_j] - Im[i_j_s];
+        Re[i_j] += Re[i_j_s];
+        Im[i_j] += Im[i_j_s];
+        Re[i_j_s] = Retmp;
+        Im[i_j_s] = Imtmp;
+      }
+    }
+
+    stepsize <<= 1;
+  }
 }
 
-// This test takes a Shepp-Logan phantom and creates a sinogram.
-BOOST_AUTO_TEST_CASE(sinogram2)
+void JPetRecoImageTools::doFFTSLOWI(std::vector<double>& Re, std::vector<double>& Im, int size, int shift)
 {
-  const auto inFile = "unitTestData/JPetRecoImageToolsTest/phantom.pgm";
-  const auto outFile = "sinogram2.ppm";
-  /// read phantom
-  std::ifstream in(inFile);
-  BOOST_REQUIRE(in);
-  std::string line;
-  getline(in, line);
-  unsigned int width;
-  unsigned int height;
-  in >> width;
-  in >> height;
-  int val;
-  in >> val; // skip max val
-  std::vector< std::vector< int > > m(width, std::vector< int >(height));
-  for (unsigned int i = 0; i < width; i++) {
-    for (unsigned int j = 0; j < height; j++) {
-      in >> val;
-      m[i][j] = val;
-    }
-  }
-  std::vector< std::vector< double > > result =
-    JPetRecoImageTools::createSinogramWithDoubleInterpolation(
-      m, 180, JPetRecoImageTools::rescale, 0, 255);
-  /// save sinogram
-  std::ofstream res(outFile);
-  res << "P2" << std::endl;
-  res << result[0].size() << " " << result.size() << std::endl;
-  res << "255" << std::endl;
-  for (unsigned int i = 0; i < result.size(); i++) {
-    for (unsigned int j = 0; j < result[0].size(); j++) {
-      res << static_cast< int >(result[i][j]) << " ";
-    }
-    res << std::endl;
-  }
-  res.close();
-  BOOST_REQUIRE(JPetCommonTools::ifFileExisting(outFile));
-}
 
-BOOST_AUTO_TEST_CASE(backProject)
-{
-  const auto inFile = "unitTestData/JPetRecoImageToolsTest/phantom.pgm";
-  const auto outFile = "backproject.ppm";
-  /// read phantom
-  std::ifstream in(inFile);
-  BOOST_REQUIRE(in);
-  std::string line;
-  getline(in, line);
-  unsigned int width;
-  unsigned int height;
-  in >> width;
-  in >> height;
-  int val;
-  in >> val; // skip max val
-  std::vector< std::vector< int > > m(width, std::vector< int >(height));
-  for (unsigned int i = 0; i < width; i++) {
-    for (unsigned int j = 0; j < height; j++) {
-      in >> val;
-      m[i][j] = val;
-    }
+  for (int i = shift; i < shift + size; ++i) {
+    Im[i] = -Im[i];
   }
-  int views = 180;
-  int scans = 256;
-  JPetRecoImageTools::Matrix2DProj sinogram =
-    JPetRecoImageTools::createSinogramWithSingleInterpolation(
-      m, views, scans, 0, 180, JPetRecoImageTools::linear,
-      JPetRecoImageTools::rescale);
-  JPetRecoImageTools::Matrix2DProj result = JPetRecoImageTools::backProject(
-        sinogram, 180, JPetRecoImageTools::rescale, 0, 255);
-  std::ofstream res(outFile);
-  res << "P2" << std::endl;
-  res << result[0].size() << " " << result.size() << std::endl;
-  res << "255" << std::endl;
-  for (unsigned int i = 0; i < result.size(); i++) {
-    for (unsigned int j = 0; j < result[0].size(); j++) {
-      res << static_cast< int >(result[i][j]) << " ";
-    }
-    res << std::endl;
-  }
-  res.close();
-  BOOST_REQUIRE(JPetCommonTools::ifFileExisting(outFile));
-}
 
-BOOST_AUTO_TEST_CASE(backProjectNone)
-{
-  const auto inFile = "unitTestData/JPetRecoImageToolsTest/phantom.pgm";
-  const auto outFile = "backprojectNone.ppm";
-  /// read phantom
-  std::ifstream in(inFile);
-  BOOST_REQUIRE(in);
-  std::string line;
-  getline(in, line);
-  unsigned int width;
-  unsigned int height;
-  in >> width;
-  in >> height;
-  int val;
-  in >> val; // skip max val
-  std::vector< std::vector< int > > m(width, std::vector< int >(height));
-  for (unsigned int i = 0; i < width; i++) {
-    for (unsigned int j = 0; j < height; j++) {
-      in >> val;
-      m[i][j] = val;
-    }
-  }
-  JPetRecoImageTools::Matrix2DProj sinogram =
-    JPetRecoImageTools::createSinogramWithDoubleInterpolation(
-      m, 180, JPetRecoImageTools::rescale, 0, 255);
-  JPetFilterNone noneFilter;
-  JPetRecoImageTools::FourierTransformFunction f = JPetRecoImageTools::doFFTW;
-  JPetRecoImageTools::Matrix2DProj filteredSinogram =
-    JPetRecoImageTools::FilterSinogram(f, noneFilter, sinogram);
-  JPetRecoImageTools::Matrix2DProj result = JPetRecoImageTools::backProject(
-        filteredSinogram, 180, JPetRecoImageTools::rescale, 0, 255);
-  std::ofstream res(outFile);
-  res << "P2" << std::endl;
-  res << result[0].size() << " " << result.size() << std::endl;
-  res << "255" << std::endl;
-  for (unsigned int i = 0; i < result.size(); i++) {
-    for (unsigned int j = 0; j < result[0].size(); j++) {
-      res << static_cast< int >(result[i][j]) << " ";
-    }
-    res << std::endl;
-  }
-  res.close();
-  BOOST_REQUIRE(JPetCommonTools::ifFileExisting(outFile));
-}
+  doFFTSLOWT(Re, Im, size, shift);
 
-BOOST_AUTO_TEST_CASE(backProjectRamLak)
-{
-  const auto inFile = "unitTestData/JPetRecoImageToolsTest/phantom.pgm";
-  const auto outFile = "backprojectRamLak.ppm";
-  /// read phantom
-  std::ifstream in(inFile);
-  BOOST_REQUIRE(in);
-  std::string line;
-  getline(in, line);
-  unsigned int width;
-  unsigned int height;
-  in >> width;
-  in >> height;
-  int val;
-  in >> val; // skip max val
-  std::vector< std::vector< int > > m(width, std::vector< int >(height));
-  for (unsigned int i = 0; i < width; i++) {
-    for (unsigned int j = 0; j < height; j++) {
-      in >> val;
-      m[i][j] = val;
-    }
+  for (int i1 = shift; i1 < shift + size; ++i1) {
+    Re[i1] /= (double)size;
+    Im[i1] = -Im[i1] / (double)size;
   }
-  JPetRecoImageTools::Matrix2DProj sinogram =
-    JPetRecoImageTools::createSinogramWithDoubleInterpolation(
-      m, 180, JPetRecoImageTools::rescale, 0, 255);
-  JPetFilterRamLak ramlakFilter;
-  JPetRecoImageTools::FourierTransformFunction f = JPetRecoImageTools::doFFTW;
-  JPetRecoImageTools::Matrix2DProj filteredSinogram =
-    JPetRecoImageTools::FilterSinogram(f, ramlakFilter, sinogram);
-  JPetRecoImageTools::Matrix2DProj result = JPetRecoImageTools::backProject(
-        filteredSinogram, 180, JPetRecoImageTools::rescale, 0, 255);
-  std::ofstream res(outFile);
-  res << "P2" << std::endl;
-  res << result[0].size() << " " << result.size() << std::endl;
-  res << "255" << std::endl;
-  for (unsigned int i = 0; i < result.size(); i++) {
-    for (unsigned int j = 0; j < result[0].size(); j++) {
-      res << static_cast< int >(result[i][j]) << " ";
-    }
-    res << std::endl;
-  }
-  res.close();
-  BOOST_REQUIRE(JPetCommonTools::ifFileExisting(outFile));
 }
-
-BOOST_AUTO_TEST_CASE(backProjectSheppLogan)
-{
-  const auto inFile = "unitTestData/JPetRecoImageToolsTest/phantom.pgm";
-  const auto outFile = "backprojectSheppLogan.ppm";
-  /// read phantom
-  std::ifstream in(inFile);
-  BOOST_REQUIRE(in);
-  std::string line;
-  getline(in, line);
-  unsigned int width;
-  unsigned int height;
-  in >> width;
-  in >> height;
-  int val;
-  in >> val; // skip max val
-  std::vector< std::vector< int > > m(width, std::vector< int >(height));
-  for (unsigned int i = 0; i < width; i++) {
-    for (unsigned int j = 0; j < height; j++) {
-      in >> val;
-      m[i][j] = val;
-    }
-  }
-  JPetRecoImageTools::Matrix2DProj sinogram =
-    JPetRecoImageTools::createSinogramWithDoubleInterpolation(
-      m, 180, JPetRecoImageTools::rescale, 0, 255);
-  JPetFilterSheppLogan shepploganFilter;
-  JPetRecoImageTools::FourierTransformFunction f = JPetRecoImageTools::doFFTW;
-  JPetRecoImageTools::Matrix2DProj filteredSinogram =
-    JPetRecoImageTools::FilterSinogram(f, shepploganFilter, sinogram);
-  JPetRecoImageTools::Matrix2DProj result = JPetRecoImageTools::backProject(
-        filteredSinogram, 180, JPetRecoImageTools::rescale, 0, 255);
-  std::ofstream res(outFile);
-  res << "P2" << std::endl;
-  res << result[0].size() << " " << result.size() << std::endl;
-  res << "255" << std::endl;
-  for (unsigned int i = 0; i < result.size(); i++) {
-    for (unsigned int j = 0; j < result[0].size(); j++) {
-      res << static_cast< int >(result[i][j]) << " ";
-    }
-    res << std::endl;
-  }
-  res.close();
-  BOOST_REQUIRE(JPetCommonTools::ifFileExisting(outFile));
-}
-
-BOOST_AUTO_TEST_CASE(backProjectCosine)
-{
-  const auto inFile = "unitTestData/JPetRecoImageToolsTest/phantom.pgm";
-  const auto outFile = "backprojectCosine.ppm";
-  /// read phantom
-  std::ifstream in(inFile);
-  BOOST_REQUIRE(in);
-  std::string line;
-  getline(in, line);
-  unsigned int width;
-  unsigned int height;
-  in >> width;
-  in >> height;
-  int val;
-  in >> val; // skip max val
-  std::vector< std::vector< int > > m(width, std::vector< int >(height));
-  for (unsigned int i = 0; i < width; i++) {
-    for (unsigned int j = 0; j < height; j++) {
-      in >> val;
-      m[i][j] = val;
-    }
-  }
-  JPetRecoImageTools::Matrix2DProj sinogram =
-    JPetRecoImageTools::createSinogramWithDoubleInterpolation(
-      m, 180, JPetRecoImageTools::rescale, 0, 255);
-  JPetFilterCosine cosineFilter;
-  JPetRecoImageTools::FourierTransformFunction f = JPetRecoImageTools::doFFTW;
-  JPetRecoImageTools::Matrix2DProj filteredSinogram =
-    JPetRecoImageTools::FilterSinogram(f, cosineFilter, sinogram);
-  JPetRecoImageTools::Matrix2DProj result = JPetRecoImageTools::backProject(
-        filteredSinogram, 180, JPetRecoImageTools::rescale, 0, 255);
-  std::ofstream res(outFile);
-  res << "P2" << std::endl;
-  res << result[0].size() << " " << result.size() << std::endl;
-  res << "255" << std::endl;
-  for (unsigned int i = 0; i < result.size(); i++) {
-    for (unsigned int j = 0; j < result[0].size(); j++) {
-      res << static_cast< int >(result[i][j]) << " ";
-    }
-    res << std::endl;
-  }
-  res.close();
-  BOOST_REQUIRE(JPetCommonTools::ifFileExisting(outFile));
-}
-
-BOOST_AUTO_TEST_CASE(backProjectHamming)
-{
-  const auto inFile = "unitTestData/JPetRecoImageToolsTest/phantom.pgm";
-  const auto outFile = "backprojectHamming.ppm";
-  /// read phantom
-  std::ifstream in(inFile);
-  BOOST_REQUIRE(in);
-  std::string line;
-  getline(in, line);
-  unsigned int width;
-  unsigned int height;
-  in >> width;
-  in >> height;
-  int val;
-  in >> val; // skip max val
-  std::vector< std::vector< int > > m(width, std::vector< int >(height));
-  for (unsigned int i = 0; i < width; i++) {
-    for (unsigned int j = 0; j < height; j++) {
-      in >> val;
-      m[i][j] = val;
-    }
-  }
-  JPetRecoImageTools::Matrix2DProj sinogram =
-    JPetRecoImageTools::createSinogramWithDoubleInterpolation(
-      m, 180, JPetRecoImageTools::rescale, 0, 255);
-  JPetFilterHamming hammingFilter(0.45);
-  JPetRecoImageTools::FourierTransformFunction f = JPetRecoImageTools::doFFTW;
-  JPetRecoImageTools::Matrix2DProj filteredSinogram =
-    JPetRecoImageTools::FilterSinogram(f, hammingFilter, sinogram);
-  JPetRecoImageTools::Matrix2DProj result = JPetRecoImageTools::backProject(
-        filteredSinogram, 180, JPetRecoImageTools::rescale, 0, 255);
-  std::ofstream res(outFile);
-  res << "P2" << std::endl;
-  res << result[0].size() << " " << result.size() << std::endl;
-  res << "255" << std::endl;
-  for (unsigned int i = 0; i < result.size(); i++) {
-    for (unsigned int j = 0; j < result[0].size(); j++) {
-      res << static_cast< int >(result[i][j]) << " ";
-    }
-    res << std::endl;
-  }
-  res.close();
-  BOOST_REQUIRE(JPetCommonTools::ifFileExisting(outFile));
-}
-
-BOOST_AUTO_TEST_CASE(backProjectRidgelet)
-{
-  const auto inFile = "unitTestData/JPetRecoImageToolsTest/phantom.pgm";
-  const auto outFile = "backprojectRidgelet.ppm";
-  /// read phantom
-  std::ifstream in(inFile);
-  BOOST_REQUIRE(in);
-  std::string line;
-  getline(in, line);
-  unsigned int width;
-  unsigned int height;
-  in >> width;
-  in >> height;
-  int val;
-  in >> val; // skip max val
-  std::vector< std::vector< int > > m(width, std::vector< int >(height));
-  for (unsigned int i = 0; i < width; i++) {
-    for (unsigned int j = 0; j < height; j++) {
-      in >> val;
-      m[i][j] = val;
-    }
-  }
-  JPetRecoImageTools::Matrix2DProj sinogram =
-    JPetRecoImageTools::createSinogramWithDoubleInterpolation(
-      m, 180, JPetRecoImageTools::rescale, 0, 255);
-  JPetFilterRidgelet ridgeletFilter;
-  JPetRecoImageTools::FourierTransformFunction f = JPetRecoImageTools::doFFTW;
-  JPetRecoImageTools::Matrix2DProj filteredSinogram =
-    JPetRecoImageTools::FilterSinogram(f, ridgeletFilter, sinogram);
-  JPetRecoImageTools::Matrix2DProj result = JPetRecoImageTools::backProject(
-        filteredSinogram, 180, JPetRecoImageTools::rescale, 0, 255);
-  std::ofstream res(outFile);
-  res << "P2" << std::endl;
-  res << result[0].size() << " " << result.size() << std::endl;
-  res << "255" << std::endl;
-  for (unsigned int i = 0; i < result.size(); i++) {
-    for (unsigned int j = 0; j < result[0].size(); j++) {
-      res << static_cast< int >(result[i][j]) << " ";
-    }
-    res << std::endl;
-  }
-  res.close();
-  BOOST_REQUIRE(JPetCommonTools::ifFileExisting(outFile));
-}
-
-BOOST_AUTO_TEST_CASE(backProjectNoneSlow)
-{
-  const auto inFile = "unitTestData/JPetRecoImageToolsTest/phantom.pgm";
-  const auto outFile = "backprojectNoneSlow.ppm";
-  /// read phantom
-  std::ifstream in(inFile);
-  BOOST_REQUIRE(in);
-  std::string line;
-  getline(in, line);
-  unsigned int width;
-  unsigned int height;
-  in >> width;
-  in >> height;
-  int val;
-  in >> val; // skip max val
-  std::vector< std::vector< int > > m(width, std::vector< int >(height));
-  for (unsigned int i = 0; i < width; i++) {
-    for (unsigned int j = 0; j < height; j++) {
-      in >> val;
-      m[i][j] = val;
-    }
-  }
-  JPetRecoImageTools::Matrix2DProj sinogram =
-    JPetRecoImageTools::createSinogramWithDoubleInterpolation(
-      m, 180, JPetRecoImageTools::rescale, 0, 255);
-  JPetFilterNone noneFilter;
-  JPetRecoImageTools::FourierTransformFunction f =
-    JPetRecoImageTools::doFFTSLOW;
-  JPetRecoImageTools::Matrix2DProj filteredSinogram =
-    JPetRecoImageTools::FilterSinogram(f, noneFilter, sinogram);
-  JPetRecoImageTools::Matrix2DProj result = JPetRecoImageTools::backProject(
-        filteredSinogram, 180, JPetRecoImageTools::rescale, 0, 255);
-  std::ofstream res(outFile);
-  res << "P2" << std::endl;
-  res << result[0].size() << " " << result.size() << std::endl;
-  res << "255" << std::endl;
-  for (unsigned int i = 0; i < result.size(); i++) {
-    for (unsigned int j = 0; j < result[0].size(); j++) {
-      res << static_cast< int >(result[i][j]) << " ";
-    }
-    res << std::endl;
-  }
-  res.close();
-  BOOST_REQUIRE(JPetCommonTools::ifFileExisting(outFile));
-}
-
-BOOST_AUTO_TEST_CASE(backProjectRamLakSlow)
-{
-  const auto inFile = "unitTestData/JPetRecoImageToolsTest/phantom.pgm";
-  const auto outFile = "backprojectRamLakSlow.ppm";
-  /// read phantom
-  std::ifstream in(inFile);
-  BOOST_REQUIRE(in);
-  std::string line;
-  getline(in, line);
-  unsigned int width;
-  unsigned int height;
-  in >> width;
-  in >> height;
-  int val;
-  in >> val; // skip max val
-  std::vector< std::vector< int > > m(width, std::vector< int >(height));
-  for (unsigned int i = 0; i < width; i++) {
-    for (unsigned int j = 0; j < height; j++) {
-      in >> val;
-      m[i][j] = val;
-    }
-  }
-  JPetRecoImageTools::Matrix2DProj sinogram =
-    JPetRecoImageTools::createSinogramWithDoubleInterpolation(
-      m, 180, JPetRecoImageTools::rescale, 0, 255);
-  JPetFilterRamLak ramlak;
-  JPetRecoImageTools::FourierTransformFunction f =
-    JPetRecoImageTools::doFFTSLOW;
-  JPetRecoImageTools::Matrix2DProj filteredSinogram =
-    JPetRecoImageTools::FilterSinogram(f, ramlak, sinogram);
-  JPetRecoImageTools::Matrix2DProj result = JPetRecoImageTools::backProject(
-        filteredSinogram, 180, JPetRecoImageTools::rescale, 0, 255);
-  std::ofstream res(outFile);
-  res << "P2" << std::endl;
-  res << result[0].size() << " " << result.size() << std::endl;
-  res << "255" << std::endl;
-  for (unsigned int i = 0; i < result.size(); i++) {
-    for (unsigned int j = 0; j < result[0].size(); j++) {
-      res << static_cast< int >(result[i][j]) << " ";
-    }
-    res << std::endl;
-  }
-  res.close();
-  BOOST_REQUIRE(JPetCommonTools::ifFileExisting(outFile));
-}
-
-BOOST_AUTO_TEST_CASE(backProjectSheppLoganSlow)
-{
-  const auto inFile = "unitTestData/JPetRecoImageToolsTest/phantom.pgm";
-  const auto outFile = "backprojectSheppLoganSlow.ppm";
-  /// read phantom
-  std::ifstream in(inFile);
-  BOOST_REQUIRE(in);
-  std::string line;
-  getline(in, line);
-  unsigned int width;
-  unsigned int height;
-  in >> width;
-  in >> height;
-  int val;
-  in >> val; // skip max val
-  std::vector< std::vector< int > > m(width, std::vector< int >(height));
-  for (unsigned int i = 0; i < width; i++) {
-    for (unsigned int j = 0; j < height; j++) {
-      in >> val;
-      m[i][j] = val;
-    }
-  }
-  JPetRecoImageTools::Matrix2DProj sinogram =
-    JPetRecoImageTools::createSinogramWithDoubleInterpolation(
-      m, 180, JPetRecoImageTools::rescale, 0, 255);
-  JPetFilterSheppLogan shepplogan;
-  JPetRecoImageTools::FourierTransformFunction f =
-    JPetRecoImageTools::doFFTSLOW;
-  JPetRecoImageTools::Matrix2DProj filteredSinogram =
-    JPetRecoImageTools::FilterSinogram(f, shepplogan, sinogram);
-  JPetRecoImageTools::Matrix2DProj result = JPetRecoImageTools::backProject(
-        filteredSinogram, 180, JPetRecoImageTools::rescale, 0, 255);
-  std::ofstream res(outFile);
-  res << "P2" << std::endl;
-  res << result[0].size() << " " << result.size() << std::endl;
-  res << "255" << std::endl;
-  for (unsigned int i = 0; i < result.size(); i++) {
-    for (unsigned int j = 0; j < result[0].size(); j++) {
-      res << static_cast< int >(result[i][j]) << " ";
-    }
-    res << std::endl;
-  }
-  res.close();
-  BOOST_REQUIRE(JPetCommonTools::ifFileExisting(outFile));
-}
-
-BOOST_AUTO_TEST_CASE(backProjectCosineSlow)
-{
-  const auto inFile = "unitTestData/JPetRecoImageToolsTest/phantom.pgm";
-  const auto outFile = "backprojectCosineSlow.ppm";
-  /// read phantom
-  std::ifstream in(inFile);
-  BOOST_REQUIRE(in);
-  std::string line;
-  getline(in, line);
-  unsigned int width;
-  unsigned int height;
-  in >> width;
-  in >> height;
-  int val;
-  in >> val; // skip max val
-  std::vector< std::vector< int > > m(width, std::vector< int >(height));
-  for (unsigned int i = 0; i < width; i++) {
-    for (unsigned int j = 0; j < height; j++) {
-      in >> val;
-      m[i][j] = val;
-    }
-  }
-  JPetRecoImageTools::Matrix2DProj sinogram =
-    JPetRecoImageTools::createSinogramWithDoubleInterpolation(
-      m, 180, JPetRecoImageTools::rescale, 0, 255);
-  JPetFilterCosine cosine;
-  JPetRecoImageTools::FourierTransformFunction f =
-    JPetRecoImageTools::doFFTSLOW;
-  JPetRecoImageTools::Matrix2DProj filteredSinogram =
-    JPetRecoImageTools::FilterSinogram(f, cosine, sinogram);
-  JPetRecoImageTools::Matrix2DProj result = JPetRecoImageTools::backProject(
-        filteredSinogram, 180, JPetRecoImageTools::rescale, 0, 255);
-  std::ofstream res(outFile);
-  res << "P2" << std::endl;
-  res << result[0].size() << " " << result.size() << std::endl;
-  res << "255" << std::endl;
-  for (unsigned int i = 0; i < result.size(); i++) {
-    for (unsigned int j = 0; j < result[0].size(); j++) {
-      res << static_cast< int >(result[i][j]) << " ";
-    }
-    res << std::endl;
-  }
-  res.close();
-  BOOST_REQUIRE(JPetCommonTools::ifFileExisting(outFile));
-}
-
-BOOST_AUTO_TEST_CASE(backProjectHammingSlow)
-{
-  const auto inFile = "unitTestData/JPetRecoImageToolsTest/phantom.pgm";
-  const auto outFile = "backprojectHammingSlow.ppm";
-  /// read phantom
-  std::ifstream in(inFile);
-  BOOST_REQUIRE(in);
-  std::string line;
-  getline(in, line);
-  unsigned int width;
-  unsigned int height;
-  in >> width;
-  in >> height;
-  int val;
-  in >> val; // skip max val
-  std::vector< std::vector< int > > m(width, std::vector< int >(height));
-  for (unsigned int i = 0; i < width; i++) {
-    for (unsigned int j = 0; j < height; j++) {
-      in >> val;
-      m[i][j] = val;
-    }
-  }
-  JPetRecoImageTools::Matrix2DProj sinogram =
-    JPetRecoImageTools::createSinogramWithDoubleInterpolation(
-      m, 180, JPetRecoImageTools::rescale, 0, 255);
-  JPetFilterHamming hamming;
-  JPetRecoImageTools::FourierTransformFunction f =
-    JPetRecoImageTools::doFFTSLOW;
-  JPetRecoImageTools::Matrix2DProj filteredSinogram =
-    JPetRecoImageTools::FilterSinogram(f, hamming, sinogram);
-  JPetRecoImageTools::Matrix2DProj result = JPetRecoImageTools::backProject(
-        filteredSinogram, 180, JPetRecoImageTools::rescale, 0, 255);
-  std::ofstream res(outFile);
-  res << "P2" << std::endl;
-  res << result[0].size() << " " << result.size() << std::endl;
-  res << "255" << std::endl;
-  for (unsigned int i = 0; i < result.size(); i++) {
-    for (unsigned int j = 0; j < result[0].size(); j++) {
-      res << static_cast< int >(result[i][j]) << " ";
-    }
-    res << std::endl;
-  }
-  res.close();
-  BOOST_REQUIRE(JPetCommonTools::ifFileExisting(outFile));
-}
-
-BOOST_AUTO_TEST_CASE(backProjectRidgeletSlow)
-{
-  const auto inFile = "unitTestData/JPetRecoImageToolsTest/phantom.pgm";
-  const auto outFile = "backprojectRidgeletSlow.ppm";
-  /// read phantom
-  std::ifstream in(inFile);
-  BOOST_REQUIRE(in);
-  std::string line;
-  getline(in, line);
-  unsigned int width;
-  unsigned int height;
-  in >> width;
-  in >> height;
-  int val;
-  in >> val; // skip max val
-  std::vector< std::vector< int > > m(width, std::vector< int >(height));
-  for (unsigned int i = 0; i < width; i++) {
-    for (unsigned int j = 0; j < height; j++) {
-      in >> val;
-      m[i][j] = val;
-    }
-  }
-  JPetRecoImageTools::Matrix2DProj sinogram =
-    JPetRecoImageTools::createSinogramWithDoubleInterpolation(
-      m, 180, JPetRecoImageTools::rescale, 0, 255);
-  JPetFilterRidgelet ridgelet;
-  JPetRecoImageTools::FourierTransformFunction f =
-    JPetRecoImageTools::doFFTSLOW;
-  JPetRecoImageTools::Matrix2DProj filteredSinogram =
-    JPetRecoImageTools::FilterSinogram(f, ridgelet, sinogram);
-  JPetRecoImageTools::Matrix2DProj result = JPetRecoImageTools::backProject(
-        filteredSinogram, 180, JPetRecoImageTools::rescale, 0, 255);
-  std::ofstream res(outFile);
-  res << "P2" << std::endl;
-  res << result[0].size() << " " << result.size() << std::endl;
-  res << "255" << std::endl;
-  for (unsigned int i = 0; i < result.size(); i++) {
-    for (unsigned int j = 0; j < result[0].size(); j++) {
-      res << static_cast< int >(result[i][j]) << " ";
-    }
-    res << std::endl;
-  }
-  res.close();
-  BOOST_REQUIRE(JPetCommonTools::ifFileExisting(outFile));
-}
-
-BOOST_AUTO_TEST_CASE(backProjectSinogramNone)
-{
-  const auto inFile = "unitTestData/JPetRecoImageToolsTest/sinogramBackproject.ppm";
-  const auto outFile = "backprojectSinogramNone.ppm";
-  /// read phantom
-  std::ifstream in(inFile);
-  BOOST_REQUIRE(in);
-  std::string line;
-  getline(in, line);
-  unsigned int width;
-  unsigned int height;
-  in >> width;
-  in >> height;
-  int val;
-  in >> val; // skip max val
-  std::vector<std::vector<double>> sinogram(height, std::vector<double>(width));
-  for (unsigned int i = 0; i < height; i++) {
-    for (unsigned int j = 0; j < width; j++) {
-      in >> val;
-      sinogram[i][j] = val;
-    }
-  }
-  JPetFilterNone noneFilter;
-  JPetRecoImageTools::FourierTransformFunction f = JPetRecoImageTools::doFFTW;
-  JPetRecoImageTools::Matrix2DProj filteredSinogram =
-    JPetRecoImageTools::FilterSinogram(f, noneFilter, sinogram);
-  JPetRecoImageTools::Matrix2DProj result = JPetRecoImageTools::backProject(
-        filteredSinogram, 360, JPetRecoImageTools::nonRescale, 0, 255);
-  int maxValue = 0;
-  for (unsigned int i = 0; i < result.size(); i++) {
-    for (unsigned int j = 0; j < result[0].size(); j++) {
-      if (static_cast<int>(result[i][j]) > maxValue)
-        maxValue = static_cast<int>(result[i][j]);
-    }
-  }
-  std::ofstream res(outFile);
-  res << "P2" << std::endl;
-  res << result[0].size() << " " << result.size() << std::endl;
-  res << maxValue << std::endl;
-  for (unsigned int i = 0; i < result.size(); i++) {
-    for (unsigned int j = 0; j < result[0].size(); j++) {
-      res << static_cast<int>(result[i][j]) << " ";
-    }
-    res << std::endl;
-  }
-  res.close();
-  BOOST_REQUIRE(JPetCommonTools::ifFileExisting(outFile));
-}
-
-BOOST_AUTO_TEST_CASE(backProjectSinogramRamLak)
-{
-  const auto inFile = "unitTestData/JPetRecoImageToolsTest/sinogramBackproject.ppm";
-  const auto outFile = "backprojectSinogramRamLak.ppm";
-  /// read phantom
-  std::ifstream in(inFile);
-  BOOST_REQUIRE(in);
-  std::string line;
-  getline(in, line);
-  unsigned int width;
-  unsigned int height;
-  in >> width;
-  in >> height;
-  int val;
-  in >> val; // skip max val
-  std::vector<std::vector<double>> sinogram(height, std::vector<double>(width));
-  for (unsigned int i = 0; i < height; i++) {
-    for (unsigned int j = 0; j < width; j++) {
-      in >> val;
-      sinogram[i][j] = val;
-    }
-  }
-  JPetFilterRamLak ramLakFilter;
-  JPetRecoImageTools::FourierTransformFunction f = JPetRecoImageTools::doFFTW;
-  JPetRecoImageTools::Matrix2DProj filteredSinogram =
-    JPetRecoImageTools::FilterSinogram(f, ramLakFilter, sinogram);
-  JPetRecoImageTools::Matrix2DProj result = JPetRecoImageTools::backProject(
-        filteredSinogram, 360, JPetRecoImageTools::nonRescale, 0, 255);
-  int maxValue = 0;
-  for (unsigned int i = 0; i < result.size(); i++) {
-    for (unsigned int j = 0; j < result[0].size(); j++) {
-      if (static_cast<int>(result[i][j]) > maxValue)
-        maxValue = static_cast<int>(result[i][j]);
-    }
-  }
-  std::ofstream res(outFile);
-  res << "P2" << std::endl;
-  res << result[0].size() << " " << result.size() << std::endl;
-  res << maxValue << std::endl;
-  for (unsigned int i = 0; i < result.size(); i++) {
-    for (unsigned int j = 0; j < result[0].size(); j++) {
-      res << static_cast<int>(result[i][j]) << " ";
-    }
-    res << std::endl;
-  }
-  res.close();
-  BOOST_REQUIRE(JPetCommonTools::ifFileExisting(outFile));
-}
-
-BOOST_AUTO_TEST_SUITE_END()
